==> 01_3T3L1_phos/02__QC.Rmd <==
---
title: "02__QC__1"
output: html_document
---

Background

Here I do QC:
Missingness
PCA
Cluster dendogram
Corrplot



###Initialise

Packages
```{r}
library(ggplot2)
library(reshape2)
library(corrplot)
library(dendextend)
source("..\\scripts/3T3L1_mouse_IR__DF__202104_objects.R")
source("..\\scripts/essential_visualisations_JvG.R")
```


###Load in data
```{r}
load("data/intermediate/phos_3t3/phos_3t3_filt_norm__1.RData")
load("data/intermediate/phos_3t3/phos_3t3_raw__1.RData")
```


###Missingness

Exact numbers in raw data
```{r}
library(tidyverse)
#Prepare
sample_quant_df <- data.frame("name" = colnames(phos_3t3_raw)[1:52],
                              "quant" = colSums(!is.na(phos_3t3_raw[, 1:52]))) %>%
  mutate(model = sapply(.[, "name"], function(x) strsplit(x, "_")[[1]][1])) %>%
  rbind(c("Total_INS", nrow(phos_3t3_raw), "Total")) %>%
  mutate(model = factor(model, levels = unique(model))) %>%
  mutate(name = factor(name, levels = unique(name))) %>%
  mutate(quant = as.numeric(quant))

#plot
ggplot() +
  geom_col(data = sample_quant_df,
           aes(x = name,
               y = quant,
               colour = model),
           fill = "white",
           size = 0.25,
           width = 0.7) +
  geom_col(data = sample_quant_df[grep("INS",
                                              sample_quant_df$name), ],
           aes(x = name,
               y = quant,
               fill = model),
           alpha = 0.4) +
  scale_colour_manual(values = c(SH_control_model_main_colours, "#bfbebe"),
                      name = "Condition") +
  scale_fill_manual(values = c(SH_control_model_main_colours, "#bfbebe"),
                    name = "Condition") +
  comfy_theme() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5, size = 7, colour = "black")) +
  expand_limits(y = 0) +
  labs(x = "Sample",
       y = "# phosphopeptides") +
  ggtitle("3T3-L1 Phosphoproteome Missingness")
ggsave("output/images/QC/bplot_3t3l1_phospho_quant.pdf",
       width = 4.25, height = 2)
write.csv(sample_quant_df,
          file = "output/data/QC/sample_quant_df.csv")
```



###Clustering and corr plot
```{r}
###Correlation heatmap
corr_m <- cor(phos_3t3_filt_norm[, 1:52], use = "pairwise.complete.obs")
pdf("output/images/QC/3T3L1_phospho__corr_heatmap.pdf")
corrplot(corr_m,
         method = "color",
         type = "upper",
         order = "hclust",
         tl.cex = 0.5,
         col = c(rep("black", 700),
                 colorRampPalette(c("white", "blue"))(100)),
         title = "3T3-L1 Phosphoproteome Corr Heatmap")
dev.off()
write.csv(corr_m,
          file = "output/data/QC/corr_m.csv")

##Min and max corr
min(corr_m)
max(corr_m[which(corr_m != 1)])
mean(corr_m[which(corr_m != 1)])


###PCA
pca <- prcomp(t(na.omit(phos_3t3_filt_norm[, 1:52])), center = TRUE, scale. = FALSE)
summary(pca)
pca_df <- as.data.frame(pca$x[, 1:2], stringsAsFactors = FALSE)
pca_df$Treatment <- sub("_\\d+",
                        "",
                        rownames(pca_df))
pca_df$Name <- rownames(pca_df)


##With colouring

#Make model factor
pca_df$Model <- sapply(pca_df$Treatment,
                       function(x) strsplit(x, "_")[[1]][1])
pca_df$Model <- factor(pca_df$Model,
                           levels = c("CTRL",
                                      "CI",
                                      "DEX",
                                      "TNF",
                                      "MPQ",
                                      "AA"))

#Plot
pca_plot <- ggplot() + 
  geom_point(data = pca_df,
             aes(x = PC1,
                 y = PC2,
                 color = Model),
             size = 3,
             shape = 21,
             stroke = 1,
             alpha = 1) +
  geom_point(data = pca_df[grep("INS",
                                pca_df$Treatment), ],
             aes(x = PC1,
                 y = PC2,
                 color = Model),
             size = 3,
             alpha = 0.4) +
  scale_colour_manual(values = SH_control_model_main_colours) + 
  labs(x = paste(c("PC1 (",
                   100*summary(pca)$importance[2, 1],
                   "%)"),
                 collapse = ""),
       y = paste(c("PC2 (",
                   100*summary(pca)$importance[2, 2],
                   "%)"),
                 collapse = "")) +
  theme(panel.background = element_blank(),
        panel.grid = element_blank(),
        axis.line = element_line(colour = "black"),
        plot.title = element_text(hjust = 0.5))
plot(pca_plot) +
  ggtitle("3T3-L1 Phosphoproteome PCA") 
  ggsave("output/images/QC/3T3L1_phospho__PCA.pdf",
         width = 5,
         height = 5)
write.csv(pca_df, file = "output/data/QC/PCA.csv")


###Clustered dendogram
dist_m <- dist(t(phos_3t3_filt_norm[, 1:52]))
h_clust <- hclust(dist_m)
h_dend <- as.dendrogram(h_clust)


h_dend_cols <- c(rep("#3F403F",
                     12),
                rep("#D73127",
                    8),
                rep("#F68C5B",
                    8),
                rep("#FED318",
                    8),
                rep("#93CA60",
                    8),
                rep("#1A984F",
                    8))
h_dend_cols <- h_dend_cols[order.dendrogram(h_dend)]
labels_colors(h_dend) <- h_dend_cols
labels_cex(h_dend) <- 0.6

pdf("output/images/QC/3T3L1_phospho__dendogram.pdf")
plot(h_dend,
     main = "3T3-L1 Phosphoproteome Dendogram")
dev.off()
```
