==> 01_3T3L1_phos/03__imputation.Rmd <==
---
title: "03__imputation__1"
output: html_document
---



Background

Here I do imputation



###Initialise

Packages
```{r}
library(ggplot2)
library(reshape2)
```


###Load in data
```{r}
load("data/intermediate/phos_3t3/phos_3t3_filt_norm__1.RData")
```


####Assess how much imputation would be done

Set up
```{r}
####Criteria: 0 in one condition, >= 5 in CTRL other condition or 4 in model other condition

###Check with imputation code
block_imputer_toimputeonly <- function(data, treatments, n_s, high_cap, low_cap){

###Create to_impute_m:
to_impute_m <- matrix(FALSE, ncol = length(treatments), nrow = nrow(data))
colnames(to_impute_m) <- treatments
rownames(to_impute_m) <- rownames(data)

na_sums_m <- matrix(0, ncol = length(treatments), nrow = nrow(data))
colnames(na_sums_m) <- treatments
rownames(na_sums_m) <- rownames(data)

data_is_na <- is.na(data)
for (k in 1:length(treatments)){
    na_sums_m[,k] <- rowSums(data_is_na[, (n_s*(k-1) + 1):(n_s*k)])
  }


##Loop over blocks:
for (k in 1:length(treatments)){

  l_inx <- n_s*(k-1) + 1
  r_inx <- n_s*k

  if (k %% 2 == 1){
    check_l_or_r <- 1
  } else {
    check_l_or_r <- -1
  }
  
  ##Loop over proteins:
  for (i in 1:nrow(data)){
    #Check <= low_cap (Note reversal of inequalities):
    if (na_sums_m[i, k] >= low_cap){
      
      #Check to left/right, >= high_cap
      if (na_sums_m[i, k + check_l_or_r] <= high_cap){
      
        counter_mean <- mean(as.numeric(data[i, (l_inx + n_s*check_l_or_r):(r_inx + n_s*check_l_or_r)]), na.rm = TRUE)
        if (na_sums_m[i, k] == n_s){
          block_mean <- counter_mean - 1
        } else {
              block_mean <- mean(as.numeric(data[i, l_inx:r_inx]), na.rm = TRUE)
        }
       
        #Check block_mean lower than counter_mean
        if (block_mean < counter_mean){
          
          #Assign to to_impute_m
          to_impute_m[i, k] <- TRUE
      }
    } 
  }
  }
  
}

return(to_impute_m)
}

phos_3t3_filt_norm_toblockimpute <- 
  cbind(block_imputer_toimputeonly(data = phos_3t3_filt_norm[, 1:12], 
                           treatments = c("CTRL_BAS", "CTRL_INS"), 
                           n_s = 6, 
                           high_cap = 1, 
                           low_cap = 6),
        block_imputer_toimputeonly(data = phos_3t3_filt_norm[, 13:52], 
                           treatments = c("CI_BAS", 
                                          "CI_INS",
                                          "DEX_BAS", 
                                          "DEX_INS",
                                          "TNF_BAS", 
                                          "TNF_INS",
                                          "MPQ_BAS", 
                                          "MPQ_INS",
                                          "AA_BAS", 
                                          "AA_INS"), 
                           n_s = 4, 
                           high_cap = 0, 
                           low_cap = 4))


###Numbers:

##In each condition:
colSums(phos_3t3_filt_norm_toblockimpute)

##Total number of sites (at least one condition)
length(which(rowSums(phos_3t3_filt_norm_toblockimpute) >= 1))

##Total number of sites into BAS
length(which(rowSums(phos_3t3_filt_norm_toblockimpute[, grep("BAS",
                                                             colnames(phos_3t3_filt_norm_toblockimpute))]) >= 1))

##Total nubmer of sites into INS
length(which(rowSums(phos_3t3_filt_norm_toblockimpute[, grep("INS",
                                                             colnames(phos_3t3_filt_norm_toblockimpute))]) >= 1))

##The one site getting imputation in basal and insulin:
phos_3t3_filt_norm[intersect(which(rowSums(phos_3t3_filt_norm_toblockimpute[, grep("BAS",
                                                             colnames(phos_3t3_filt_norm_toblockimpute))]) >= 1),
          which(rowSums(phos_3t3_filt_norm_toblockimpute[, grep("INS",
                                                             colnames(phos_3t3_filt_norm_toblockimpute))]) >= 1)),]
```


####Imputation


Impute below lowest value
```{r}
###Impute by downshifting from the lowest quantified value
#This is Sean's idea
#Sample from N(lowest_value - down_shift, SD), where lowest_value is the lowest value for the whole ppeptide, and SD is the SD of that ppeptide in the quantified group. E.g. if I am to impute into CTRL_BAS, I take the SD of CTRL_INS for that pppeptide
#to_impute_m is boolean matrix where colnames are conditions (e.g. CTRL_BAS, CTRL_INS), and TRUE indicates that a conditon needs to be imputed into
#data is data (numeric only i.e. no extra columns)
block_imp_below_site_min <- function(to_impute_m,
                                     data,
                                     downshift = 2){
    
    ##Get n structure (number of ns in each condition)
    conditions <- colnames(to_impute_m)
    n_structure <- sapply(conditions,
                              function(x) length(grep(x,
                                                      colnames(data))))
    
    ##Calculate site SDs in each condition, and site mins across conditions
    site_mins <- apply(data,
                       1,
                       FUN = "min",
                       na.rm = TRUE)
    condition_site_SDs <- matrix(0,
                            ncol = length(conditions),
                            nrow = nrow(data))
    colnames(condition_site_SDs) <- conditions
    rownames(condition_site_SDs) <- rownames(data)
    
    for (j in 1:length(conditions)){
      
      condition_site_SDs[, j] <- apply(data[, grep(conditions[j],
                                              colnames(data))],
                                  1,
                                  FUN = "sd",
                                  na.rm = TRUE)
    }
    
    
    ##Make start index. This will be used once I loop over individual values
    start_index <- 0
    
    
    
    ##Impute by looping over conditions:
    for (j in 1:length(conditions)){
    
    
    ##Loop over rows:
    for (i in 1:nrow(data)){
      
      if (to_impute_m[i, j]){
        
        ##Loop over each value within that condition
        for (k in 1:n_structure[j]){
          
          ##Check if value missing
          if (is.na(data[i, start_index + k])){
            
            data[i, start_index + k] <- rnorm(1, 
                                              mean = site_mins[i] - downshift,
                                              sd = condition_site_SDs[i, j + (-1)^(1 + j %% 2)])
          }
        }
      }
    }
    
    #Update start index
    start_index <- start_index + n_structure[j]
    }
    
    ##Return
    return(data)
}



###Impute with downshift 1 
phos_3t3_filt_norm_imp_below_ds1 <- phos_3t3_filt_norm
phos_3t3_filt_norm_imp_below_ds1[, 1:52] <- block_imp_below_site_min(to_impute_m = phos_3t3_filt_norm_toblockimpute,
                                                                     data = phos_3t3_filt_norm[, 1:52],
                                                                     downshift = 1)
phos_3t3_filt_norm_imp_below_ds1[, sapply(colnames(phos_3t3_filt_norm_toblockimpute),
                                          function(x) paste(c(x, 
                                                              "_blockimp"),
                                                            collapse = ""))] <- phos_3t3_filt_norm_toblockimpute
```



###Export
```{r}
save(phos_3t3_filt_norm_imp_below_ds1,
     file = "output/data/phos_3t3_filt_norm_imp_below_ds1__1.Rdata")
```
