==> 01_3T3L1_phos/04__stats.Rmd <==
---
title: "04__stats__2"
output: html_document
---



Background

Here I do stats



###Initialise

Packages
```{r}
library(reshape2)
library(multcomp)
```


###Load in data
```{r}
load("data/intermediate/phos_3t3/phos_3t3_filt_norm_imp_below_ds1__1.RData")

phos_3t3_filt_norm_imp_below_ds1_looser2stats <- phos_3t3_filt_norm_imp_below_ds1
```





###Filter before tests
```{r}
####NA in each model
conditions <- c("CTRL_BAS",
                "CTRL_INS",
                "CI_BAS",
                "CI_INS",
                "DEX_BAS",
                "DEX_INS",
                "TNF_BAS",
                "TNF_INS",
                "MPQ_BAS",
                "MPQ_INS",
                "AA_BAS",
                "AA_INS")
NA_per_condition_m <- matrix(0,
                             nrow = nrow(phos_3t3_filt_norm_imp_below_ds1_looser2stats),
                             ncol = length(conditions))
rownames(NA_per_condition_m) <- rownames(phos_3t3_filt_norm_imp_below_ds1_looser2stats)
colnames(NA_per_condition_m) <- conditions

for (j in 1:length(conditions)){
  
  NA_per_condition_m[, j] <- rowSums(is.na(phos_3t3_filt_norm_imp_below_ds1_looser2stats[, grep(paste(c(conditions[j],
                                                                                  "_\\d+"),
                                                                                collapse = ""),
                                                                          colnames(phos_3t3_filt_norm_imp_below_ds1_looser2stats))]))
}


####Filter before t-test
phos_3t3_filt_norm_imp_below_ds1_looser2stats_tfilt <- phos_3t3_filt_norm_imp_below_ds1_looser2stats[which(NA_per_condition_m[, "CTRL_BAS"] <= 4 &
                                                               NA_per_condition_m[, "CTRL_INS"] <= 4), ]




####Filter before ANOVA
#Rules: Quantified >= 2 in BAS and INS CTRL, has one model where both BAS and INS are quantified at least 2 times

##Make filtered df
phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt <- phos_3t3_filt_norm_imp_below_ds1_looser2stats
phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt$CTRL_filter <- NA_per_condition_m[, "CTRL_BAS"] <= 4 &
  NA_per_condition_m[, "CTRL_INS"] <= 4
phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt$CI_filter <- (NA_per_condition_m[, "CI_BAS"] <= 2 &
  NA_per_condition_m[, "CI_INS"] <= 2)
phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt$DEX_filter <- (NA_per_condition_m[, "DEX_BAS"] <= 2 &
  NA_per_condition_m[, "DEX_INS"] <= 2)
phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt$TNF_filter <- (NA_per_condition_m[, "TNF_BAS"] <= 2 &
  NA_per_condition_m[, "TNF_INS"] <= 2)
phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt$MPQ_filter <- (NA_per_condition_m[, "MPQ_BAS"] <= 2 &
  NA_per_condition_m[, "MPQ_INS"] <= 2)
phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt$AA_filter <- (NA_per_condition_m[, "AA_BAS"] <= 2 &
  NA_per_condition_m[, "AA_INS"] <= 2)
phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt$model_filter <- phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt$CI_filter +
  phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt$DEX_filter +
  phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt$TNF_filter +
  phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt$MPQ_filter +
  phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt$AA_filter > 0
phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt$overall_filter <- phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt$CTRL_filter +
  phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt$model_filter == 2

#Summary
length(which(phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt$overall_filter == TRUE))
nrow(phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt)

##Remove sites that don't pass filter
phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt <- phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt[phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt$overall_filter, ]
```


###FCs for ANOVA and Dunnett's
```{r}
####FCs at the sample level
###Function to get FC
###Uses medians
FC_maker <- function(data, treatment, n_s){
treatment_data <- data[, grep(treatment, colnames(data))]
medians <- apply(treatment_data[, 1:n_s], 
                 1,
                 median,
                 na.rm = TRUE)
FC_data <- sweep(treatment_data[, (n_s + 1):(2*n_s)],
                 1,
                 medians,
                 FUN = "-")
return(FC_data)
}

###Build FC data
treatments <- c("CTRL",
                "CI",
                "DEX",
                "TNF",
                "MPQ",
                "AA")
for (i in 1:length(treatments)){
  temp_treatment <- treatments[i]
  if (i == 1){
    phos_3t3_filt_norm_imp_below_ds1_looser2stats_FC_ANOVAfilt <- FC_maker(phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt, temp_treatment, 6)
  } else {
    phos_3t3_filt_norm_imp_below_ds1_looser2stats_FC_ANOVAfilt <- cbind(phos_3t3_filt_norm_imp_below_ds1_looser2stats_FC_ANOVAfilt, 
                         FC_maker(phos_3t3_filt_norm_imp_below_ds1_looser2stats_ANOVAfilt, temp_treatment, 4))
  }
}
colnames(phos_3t3_filt_norm_imp_below_ds1_looser2stats_FC_ANOVAfilt) <- sub("INS",
                             "FC",
                             colnames(phos_3t3_filt_norm_imp_below_ds1_looser2stats_FC_ANOVAfilt))
```

###t-tests for CTRL
```{r}
###run t.tests
for (i in 1:nrow(phos_3t3_filt_norm_imp_below_ds1_looser2stats_tfilt)){
  test <- try(t.test(phos_3t3_filt_norm_imp_below_ds1_looser2stats_tfilt[i, 1:6],
                 phos_3t3_filt_norm_imp_below_ds1_looser2stats_tfilt[i, 7:12],
                 alternative = "two.sided",
                 paired = FALSE),
              silent = TRUE)
  
  if (inherits(test, "try-error")){
    
    phos_3t3_filt_norm_imp_below_ds1_looser2stats_tfilt$t[i] <- NA
    phos_3t3_filt_norm_imp_below_ds1_looser2stats_tfilt$p_val[i] <- NA
    
  } else {
    
    phos_3t3_filt_norm_imp_below_ds1_looser2stats_tfilt$t[i] <- test$statistic
    phos_3t3_filt_norm_imp_below_ds1_looser2stats_tfilt$p_val[i] <- test$p.value
    
  }
  
}

phos_3t3_filt_norm_imp_below_ds1_looser2stats_tfilt$adj_p_val <- p.adjust(phos_3t3_filt_norm_imp_below_ds1_looser2stats_tfilt$p_val,
         method = "fdr")


###Summary
length(which(phos_3t3_filt_norm_imp_below_ds1_looser2stats_tfilt$adj_p_val < 0.05))
```



###ANOVAs
```{r}
####ANOVA

###Set up inputs:
treats <- c(rep("CTRL", 6),
                rep("CI", 4),
                rep("DEX", 4),
                rep("TNF", 4),
                rep("MPQ", 4),
                rep("AA", 4))
fcs <- phos_3t3_filt_norm_imp_below_ds1_looser2stats_FC_ANOVAfilt


treatment <- factor(treats, levels = unique(treats))

###Normal ANOVA

#p value
normal_anova_p <- NULL
for(i in 1:nrow(fcs)){
  
  #If no FC values for control
  if(rowSums(is.na(fcs[i,1:6])) == 6){
    normal_anova_p[i] <- NA
  
  #Else run ANOVA
  } else {
  p_val <- summary(aov(as.numeric(fcs[i,]) ~ treatment, data = fcs))[[1]][["Pr(>F)"]][1]
  
  #If NULL from ANOVA
  if (is.null(p_val)){
    normal_anova_p[i] <- NA
  } else {
  normal_anova_p[i] <- p_val
  }
}
}
names(normal_anova_p) <- rownames(fcs)



#f
normal_anova_f <- NULL
for(i in 1:nrow(fcs)){
  
  #If no FC values for control
  if(rowSums(is.na(fcs[i,1:6])) == 6){
    normal_anova_f[i] <- NA
  
  #Else run ANOVA
  } else {
  f_val <- summary(aov(as.numeric(fcs[i,]) ~ treatment, data = fcs))[[1]][["F value"]][1]
  
  #If NULL from ANOVA
  if (is.null(f_val)){
    normal_anova_f[i] <- NA
  } else {
  normal_anova_f[i] <- f_val
  }
}
}
names(normal_anova_f) <- rownames(fcs)
```

###Dunnett's

Filter before dunnets
```{r}
###Adjust p-values
normal_anova_p_adj <- p.adjust(normal_anova_p, method = "fdr")

###Filter
normal_anova_sig_names <- names(normal_anova_p_adj)[which(normal_anova_p_adj < 0.05)]
length(normal_anova_sig_names) #5550 compared to 24574 total
fcs_filt <- fcs[normal_anova_sig_names, ]

###Rerun ANOVA, extract fit object
filt_anova_fit <- list()
for(i in 1:nrow(fcs_filt)){
  filt_anova_fit[[i]] <- aov(as.numeric(fcs_filt[i,]) ~ treatment, data = fcs_filt)
}
```

Run Dunnett's
```{r}
#Dunnett's posthoc 
dunnet_models <- NULL
for(i in 1:nrow(fcs_filt)){
  dunnet_models[[i]] <- glht(filt_anova_fit[[i]], linfct = mcp(treatment = "Dunnett"))
}


###Extract p-values
dunnet_p_val_m <- matrix(NA,
                         ncol = 5,
                         nrow = nrow(fcs_filt))
colnames(dunnet_p_val_m) <- c("CI",
                              "DEX",
                              "TNF",
                              "MPQ",
                              "AA")
rownames(dunnet_p_val_m) <- rownames(fcs_filt)

for(i in 1:nrow(fcs_filt)){
  used_comparisons <- rownames(summary(dunnet_models[[i]])$linfct)
  used_treatments <- sub(" - CTRL", 
                       "",
                       used_comparisons)
  dunnet_p_val_m[i, used_treatments] <- summary(dunnet_models[[i]])$test$pvalues
}

###Adjust p-values
dunnet_p_val_m_adj <- dunnet_p_val_m
for (j in 1:ncol(dunnet_p_val_m)){
  dunnet_p_val_m_adj[, j] <- p.adjust(dunnet_p_val_m[, j],
                                      method = "fdr")
}

###Summary
length(which(dunnet_p_val_m_adj[, 1] < 0.05))
length(which(dunnet_p_val_m_adj[, 2] < 0.05))
length(which(dunnet_p_val_m_adj[, 3] < 0.05))
length(which(dunnet_p_val_m_adj[, 4] < 0.05))
length(which(dunnet_p_val_m_adj[, 5] < 0.05))
```


###Combine all
```{r}
phos_3t3_filt_norm_imp_below_ds1_looser2stats_stats <- phos_3t3_filt_norm_imp_below_ds1_looser2stats

###Add CTRL t
for (i in 1:nrow(phos_3t3_filt_norm_imp_below_ds1_looser2stats_tfilt)){
  
  temp_site <- rownames(phos_3t3_filt_norm_imp_below_ds1_looser2stats_tfilt)[i]
  phos_3t3_filt_norm_imp_below_ds1_looser2stats_stats[which(rownames(phos_3t3_filt_norm_imp_below_ds1_looser2stats_stats) ==
                                       temp_site), c("t",
                                   "t_p_val",
                                   "t_adj_p_val")] <-
    phos_3t3_filt_norm_imp_below_ds1_looser2stats_tfilt[i, c("t",
                                 "p_val",
                                 "adj_p_val")]
}

###Add ANOVA
for (i in 1:length(normal_anova_p)){
  
  temp_site <- names(normal_anova_p)[i]
  phos_3t3_filt_norm_imp_below_ds1_looser2stats_stats[which(rownames(phos_3t3_filt_norm_imp_below_ds1_looser2stats_stats) ==
                                       temp_site), c("normal_anova_p",
                                   "normal_anova_p_adj",
                                   "normal_anova_F")] <-
    c(normal_anova_p[i],
      normal_anova_p_adj[i],
      normal_anova_f[i])
}

###Add posthoc

for (i in 1:nrow(dunnet_p_val_m)){
  
  temp_site <- rownames(dunnet_p_val_m)[i]
  phos_3t3_filt_norm_imp_below_ds1_looser2stats_stats[which(rownames(phos_3t3_filt_norm_imp_below_ds1_looser2stats_stats) ==
                                       temp_site), 
                               c("CI_p",
                                 "DEX_p",
                                 "TNF_p",
                                 "MPQ_p",
                                 "AA_p")] <- dunnet_p_val_m[i, ]
  phos_3t3_filt_norm_imp_below_ds1_looser2stats_stats[which(rownames(phos_3t3_filt_norm_imp_below_ds1_looser2stats_stats) ==
                                       temp_site), 
                               c("CI_padj",
                                 "DEX_padj",
                                 "TNF_padj",
                                 "MPQ_padj",
                                 "AA_padj")] <- dunnet_p_val_m_adj[i, ]
}
```


###Export
```{r}
save(phos_3t3_filt_norm_imp_below_ds1_looser2stats_stats,
     file = "output/data/phos_3t3_filt_norm_imp_below_ds1_looser2stats_stats__1.RData")
```
