==> 01_3T3L1_phos/06__reg_defective_sites.Rmd <==
---
title: "07__reg_defective_sites__5"
output: html_document
editor_options: 
  chunk_output_type: inline
---

Background

Here I filter for regulated and defective sites and do basic summaries





###Initialise
```{r}
library(reshape2)
library(ggplot2)
source("..\\scripts/essential_visualisations_JvG.R")
source("..\\scripts/enrichment_analysis_JvG.R")
source("..\\scripts/3T3L1_mouse_IR__DF__202104_objects.R")
source("..\\scripts/3T3L1_mouse_IR__DF__202104_functions.R")
```

Load in data
```{r}
load("data/intermediate/phos_3t3/phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols__1.RData")
phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef <- phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols
```



###Filter for regulated and defective sites

Regulated
```{r}
###Filter
phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef
phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg <- "unregulated"
phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg[which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$t_adj_p_val < 0.05 &
                                                   phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_FC > 0.58)] <- "up"
phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg[which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$t_adj_p_val < 0.05 &
                                                   phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_FC < -0.58)] <- "down"

###Summarise
table(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg)
```

Defective

Rules:

I call a site defective in a given model if either:
CTRL is upregulated (adj_pval < 0.05, logFC > 0.58) and model has dunnets_adj_pval < 0.05, logFC < 0.58
OR
CTRL is downregulated (adj_pval < 0.05, logFC < -0.58) and model has dunnets_adj_pval < 0.05, logFC > -0.58
I then say a site is defective in multiple models simply by looking for sites that fit this criteria in multiple models


```{r}
###Create "model_defect" e.g. "CI_defect" columns
models <- c("CI",
            "DEX",
            "TNF",
            "MPQ",
            "AA")
for (j in 1:length(models)){
  
  ##Make colnames
  model_defect_colname <- paste(c(models[j],
                                  "_defect"),
                                collapse = "")
  model_FC_colname <- paste(c(models[j],
                              "_FC"),
                            collapse = "")
  model_padj_colname <- paste(c(models[j],
                                "_padj"),
                              collapse = "")
  
  ##Set model_defect_colname to false
  phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[, model_defect_colname] <- FALSE
  
  for (i in 1:nrow(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef)){
    
    if (phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg[i] == "up"){
      
      defect_verdict <-  phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[i, model_FC_colname] < 0.58 &
        phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[i, model_padj_colname] < 0.05
      
      if(is.na(defect_verdict) == FALSE){
        phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[i, model_defect_colname] <- defect_verdict
      }
    } else if (phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg[i] == "down"){
      
      defect_verdict <-  phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[i, model_FC_colname] > -0.58 &
        phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[i, model_padj_colname] < 0.05
      
      if(is.na(defect_verdict) == FALSE){
        phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[i, model_defect_colname] <- defect_verdict
      }
    }
  }
}


###Create "num_defective_models" column
phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$num_defective_models <- 
  rowSums(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[, c("CI_defect",
                                                "DEX_defect",
                                                "TNF_defect",
                                                "MPQ_defect",
                                                "AA_defect")])

###Create "is_defective" column (defective in at least one model)
phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$is_defective <- phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$num_defective_models >= 1
```


###Summary

Volcano plot
```{r}
volcano_plot_coloured(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef,
                                  pval_col = "t_adj_p_val",
                                  FC_col = "CTRL_FC",
                      colour_col = "CTRL_reg",
                                  colour_col_levels = c("unregulated", "up", "down"),
                                  colour_col_colours = c("unregulated" = "grey",
                                                         up_down_colours))
ggsave_pdfpng("output/images/insreg/vplot_insreg_3t31l1",
              width = 2.5, height = 2.5)
```

Full summary
```{r}
###Make summary
reg_defect_summary <- reg_sites_summariser(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef,
                                           all_def_row = TRUE)
write.csv(reg_defect_summary,
          file = "output/data/defect/summary/regdef_summary_3t3l1_DIA.csv")


###Visualise (barplots)

##Up vs down
#Exact numbers

#
reg_defect_summary_up_vs_down <- melt(reg_defect_summary,
                                       id.vars = "num_defective_models",
                                       measure.vars = c("all_models_up",
                                                        "all_models_down"))

reg_defect_summary_up_vs_down$num_defective_models <- factor(reg_defect_summary_up_vs_down$num_defective_models,
                                                             levels = c("0",
                                                                        "all_defective_sites",
                                                                        "1",
                                                                        "2",
                                                                        "3",
                                                                        "4",
                                                                        "5"))


ggplot(reg_defect_summary_up_vs_down,
       aes(x = num_defective_models,
           y = value,
           fill = variable)) + 
  geom_col(position = "stack",
           width = 0.8) +
  scale_fill_manual(name = "Insulin Regulation",
                      labels = c("Upregulated",
                                 "Downregulated"),
                      values = up_down_colours) +
  theme(panel.grid = element_blank(),
        panel.background = element_blank(),
        axis.line = element_line(color = "black"),
        plot.title = element_text(hjust = 0.5,
                                  size = 8),
        axis.text = element_text(colour = "black",
                                 size = 8),
        axis.title = element_text(colour = "black",
                                  size = 8),
        legend.text = element_text(colour = "black",
                                   size = 8),
        legend.title = element_text(colour = "black",
                                    size = 8),
        legend.key.size = unit(0.1,
                                   units = "inch")) +
  labs(x = "Number of defective treatments",
       y = "Insulin-regulated phosphopeptides") +
  ggtitle("Insulin-regulated Peptides In Vitro") 
  ggsave("output/images/defect/summary/3t3l1_defect_sites_up_down.pdf",
         width = 3,
         height = 3)





##Num sig_fc for each drug

reg_defect_summary_sig_fc_drugs <- melt(reg_defect_summary[2:6, ],
                                         id.vars = "num_defective_models",
                                         measure.vars = c("CI_total",
                                                               "DEX_total",
                                                               "TNF_total",
                                                               "MPQ_total",
                                                               "AA_total"))

#Reorder treatments 
reg_defect_summary_sig_fc_drugs$variable <- factor(reg_defect_summary_sig_fc_drugs$variable,
                                                    levels = c("CI_total",
                                                               "DEX_total",
                                                               "TNF_total",
                                                               "MPQ_total",
                                                               "AA_total"))
#Make model as a column
reg_defect_summary_sig_fc_drugs$model <- factor(gsub("_total",
                                               "",
                                               reg_defect_summary_sig_fc_drugs$variable),
                                               levels = c("CI",
                                                          "DEX",
                                                          "TNF",
                                                          "MPQ",
                                                          "AA"))
#factor for num_defective_models
reg_defect_summary_sig_fc_drugs$num_defective_models_factor <- factor(reg_defect_summary_sig_fc_drugs$num_defective_models)

#x-axis number defective models
ggplot(reg_defect_summary_sig_fc_drugs,
       aes(x = num_defective_models,
           y = value,
           fill = variable)) +
  scale_fill_manual(values = SH_model_main_colours,
                    name = "Treatment",
                    labels = c("CI",
                               "DEX",
                               "TNF",
                               "MPQ",
                               "AA")) +
  geom_col(position = "stack",
           width = 0.8) + 
  labs(x = "Number of defective treatments",
       y = "Number of peptides") +
  theme(panel.grid = element_blank(),
        panel.background = element_blank(),
                axis.line = element_line(color = "black"),
        axis.text = element_text(colour = "black"))  +
  ggtitle("Defective Peptides") 
  ggsave("output/images/defect/summary/3t3l1_defect_sites_per_treatment.pdf",
         width = 3, height = 4)


#x-axis model, alpha number of defective models
ggplot(reg_defect_summary_sig_fc_drugs,
       aes(x = model,
           y = value,
           fill = model,
           alpha = num_defective_models_factor)) +
  geom_col(position = "stack",
           width = 0.8) +
  scale_fill_manual(values = SH_model_main_colours,
                    name = "Model",
                    labels = c("CI",
                               "DEX",
                               "TNF",
                               "MPQ",
                               "AA")) +
  scale_alpha_manual(values = rev(seq(from = 0.2,
                                  to = 1,
                                  by = (1 - 0.2)/4)),
                     name = "# defective models")  +
  labs(x = "Model",
       y = "Number of peptides") +
  theme(panel.grid = element_blank(),
        panel.background = element_blank(),
                axis.line = element_line(color = "black"),
        axis.text.x = element_text(angle = 90,
                                   hjust = 1,
                                   vjust = 0.5,
                                   colour = "black"),
        axis.text.y = element_text(colour = "black")) +
  ggtitle("Defective Peptides") 
  ggsave("output/images/defect/summary/3t3l1_defect_sites_per_treatment_xaxismodel.pdf",
         width = 3.5, height = 4)
```


###Overrepresentation of dephos sites in defect
```{r}
overrep_pvals <- c()

###Defective geq 1
#Carve up regulated sites into up and down, and defective and not defective
print("def geq 1")
temp_test <- fishers_DE_pathway(background_genes = 
                     rownames(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg != "unregulated"), ]),
                   DE_genes = 
                     rownames(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$is_defective == TRUE), ]),
                   pathway_genes = 
                     rownames(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "down"), ]),
                   alternative = "greater")
overrep_pvals[1] <- temp_test$pval
overrep_pvals[1]

#Get likelihoods of defect in phos and dephos
#0.3223966 up
#0.5737705 down
print("defective probabilities")
length(which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "up" &
               phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$num_defective_models >= 1)) /
  length(which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "up"))
length(which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "down" &
               phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$num_defective_models >= 1)) /
  length(which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "down"))


###Defective geq 2
#Carve up regulated sites into up and down, and defective and not defective
print("def geq 2")
temp_test <- fishers_DE_pathway(background_genes = 
                     rownames(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg != "unregulated"), ]),
                   DE_genes = 
                     rownames(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$num_defective_models >= 2), ]),
                   pathway_genes = 
                     rownames(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "down"), ]),
                   alternative = "greater")
overrep_pvals[2] <- temp_test$pval
overrep_pvals[2]

#Get likelihoods of defect in phos and dephos
#0.02496434 up
#0.1147541 down
print("defective probabilities")
length(which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "up" &
               phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$num_defective_models >= 2)) /
  length(which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "up"))
length(which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "down" &
               phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$num_defective_models >= 2)) /
  length(which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "down"))

###Defective geq 3
#Carve up regulated sites into up and down, and defective and not defective
print("def geq 3")
temp_test <- fishers_DE_pathway(background_genes = 
                     rownames(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg != "unregulated"), ]),
                   DE_genes = 
                     rownames(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$num_defective_models >= 3), ]),
                   pathway_genes = 
                     rownames(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "down"), ]),
                   alternative = "greater")
overrep_pvals[3] <- temp_test$pval
overrep_pvals[3]

#Get likelihoods of defect in phos and dephos
#0.02496434 up
#0.1147541 down
print("defective probabilities")
length(which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "up" &
               phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$num_defective_models >= 3)) /
  length(which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "up"))
length(which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "down" &
               phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$num_defective_models >= 3)) /
  length(which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "down"))


###Defective geq 4
print("def geq 4")
#Carve up regulated sites into up and down, and defective and not defective
temp_test <- fishers_DE_pathway(background_genes = 
                     rownames(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg != "unregulated"), ]),
                   DE_genes = 
                     rownames(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$num_defective_models >= 4), ]),
                   pathway_genes = 
                     rownames(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "down"), ]),
                   alternative = "greater")
overrep_pvals[4] <- temp_test$pval
overrep_pvals[4]

#Get likelihoods of defect in phos and dephos
#0.02496434 up
#0.1147541 down
print("defective probabilities")
length(which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "up" &
               phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$num_defective_models >= 4)) /
  length(which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "up"))
length(which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "down" &
               phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$num_defective_models >= 4)) /
  length(which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "down"))

###Defective geq 5
print("def geq 5")
#Carve up regulated sites into up and down, and defective and not defective
temp_test <- fishers_DE_pathway(background_genes = 
                     rownames(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg != "unregulated"), ]),
                   DE_genes = 
                     rownames(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$num_defective_models >= 5), ]),
                   pathway_genes = 
                     rownames(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "down"), ]),
                   alternative = "greater")
overrep_pvals[5] <- temp_test$pval
overrep_pvals[5]

#Get likelihoods of defect in phos and dephos
#0.02496434 up
#0.1147541 down
print("defective probabilities")
length(which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "up" &
               phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$num_defective_models >= 5)) /
  length(which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "up"))
length(which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "down" &
               phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$num_defective_models >= 5)) /
  length(which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$CTRL_reg == "down"))


#Adjust pvals
p.adjust(overrep_pvals,
         method = "fdr")
```

###Defective in >= 4 models

Visualise
```{r}
fc_sig_heatmap(data = phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$num_defective_models >= 4), ],
               fc_cols_w_pvals = c("CTRL_FC",
                           "CI_FC",
                           "DEX_FC",
                           "TNF_FC",
                           "MPQ_FC",
                           "AA_FC"),
               pval_cols = c("t_adj_p_val",
                             "CI_padj",
                             "DEX_padj",
                             "TNF_padj",
                             "MPQ_padj",
                             "AA_padj"),
               order_column = "CTRL_FC",
               is_decreasing = FALSE,
               x_axis_names = c("CTRL",
                                "CI",
                                "DEX",
                                "TNF",
                                "MPQ",
                                "AA")) +
  ggtitle("Defective in 4 or more models") 
  ggsave("output/images/defect/3t3l1_defective_geq4_heatmap.pdf",
         width = 7, height = 7)
```

###Enrichment
Defect enricher function
```{r}
###Function that performs pathway enrichment for defective phosphosites

#Up to date 20210803
#Performs on defective sites that are up, down, and both
phospho_defective_enricher <- function(background_genes,
                                       defect_df,
                                       pathways_list,
                                       pathway_DE_intersection_threshold){
    ###Set up output list
    output_list <- list()
    
    ###Run enrichment
    
    #up and down
    output_list[[1]] <- as.data.frame(pathway_enricher_from_list(background_genes = background_genes,
                                           DE_genes = unique(defect_df$gene),
                                           pathways_list = pathways_list,
                                           pathway_DE_intersection_threshold = pathway_DE_intersection_threshold))
    
    #up
    output_list[[2]] <- as.data.frame(pathway_enricher_from_list(background_genes = background_genes,
                                           DE_genes = unique(defect_df$gene[which(defect_df$CTRL_FC > 0.58)]),
                                           pathways_list = pathways_list,
                                           pathway_DE_intersection_threshold = pathway_DE_intersection_threshold))
    #down
    output_list[[3]] <- as.data.frame(pathway_enricher_from_list(background_genes = background_genes,
                                           DE_genes = unique(defect_df$gene[which(defect_df$CTRL_FC < -0.58)]),
                                           pathways_list = pathways_list,
                                           pathway_DE_intersection_threshold = pathway_DE_intersection_threshold))
    names(output_list) <- c("up_and_down",
                            "up",
                            "down")
    return(output_list)
}


#Same, but for GO
phospho_defective_enricher_GO <- function(background_genes,
                                       defect_df,
                                       gene_label,
                                       organism_database,
                                       ontology,
                                       pathway_DE_intersection_threshold){
    ###Set up output list
    output_list <- list()
    
    ###Run enrichment
    
    #up and down
        output_list[[1]] <- as.data.frame(GO_enricher(background_genes = background_genes,
                                           DE_genes = unique(defect_df$gene),
                                           gene_label = gene_label,
                                           organism_database = organism_database,
                                           ontology = ontology,
                                           pathway_DE_intersection_threshold = pathway_DE_intersection_threshold))

    
    
    #up
        output_list[[2]] <- as.data.frame(GO_enricher(background_genes = background_genes,
                                           DE_genes = unique(defect_df$gene[which(defect_df$CTRL_FC > 0.58)]),
                                           gene_label = gene_label,
                                           organism_database = organism_database,
                                           ontology = ontology,
                                           pathway_DE_intersection_threshold = pathway_DE_intersection_threshold))
    #down
    output_list[[3]] <- as.data.frame(GO_enricher(background_genes = background_genes,
                                           DE_genes = unique(defect_df$gene[which(defect_df$CTRL_FC < -0.58)]),
                                           gene_label = gene_label,
                                           organism_database = organism_database,
                                           ontology = ontology,
                                           pathway_DE_intersection_threshold = pathway_DE_intersection_threshold))
    names(output_list) <- c("up_and_down",
                            "up",
                            "down")
    return(output_list)
}
```

GO
```{r}
###BG: All proteins with quantified ppetides

##BP
#Defective Geq 2
GO_BP_overrep_defect_geq2 <- phospho_defective_enricher_GO(background_genes  = unique(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$gene),
                            defect_df = 
                              phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$num_defective_models >= 2), ],
                            gene_label = "SYMBOL",
                            organism_database = org.Mm.eg.db,
                            ontology = "BP",
                            pathway_DE_intersection_threshold = 3) 

##MF
#Defective Geq 2
GO_MF_overrep_defect_geq2 <- phospho_defective_enricher_GO(background_genes  = unique(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$gene),
                            defect_df = 
                              phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$num_defective_models >= 2), ],
                            gene_label = "SYMBOL",
                            organism_database = org.Mm.eg.db,
                            ontology = "MF",
                            pathway_DE_intersection_threshold = 3) 

##CC
#Defective Geq 2
GO_CC_overrep_defect_geq2 <- phospho_defective_enricher_GO(background_genes  = unique(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$gene),
                            defect_df = 
                              phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[which(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef$num_defective_models >= 2), ],
                            gene_label = "SYMBOL",
                            organism_database = org.Mm.eg.db,
                            ontology = "CC",
                            pathway_DE_intersection_threshold = 3) 
  

###Visualise
##Geq 2

#BP
pathway_enrichment_barplot(GO_BP_overrep_defect_geq2[[1]],
                           num_pathways = "sig_only",
                           sig_line = TRUE,
                           sig_cutoff = 0.05,
                           colour = "grey") + 
  ggtitle("Biological processes with dysregulation in 2 or more models") 
  ggsave("output/images/defect/enrichment/GO/GOBP_defectgeq2.pdf",
         width = 7,
         height = 3)
pathway_enrichment_barplot(GO_BP_overrep_defect_geq2[[2]],
                           num_pathways = "sig_only",
                           sig_line = TRUE,
                           sig_cutoff = 0.05,
                           colour = "red")+ 
  ggtitle("Biological processes with phosphorylation dysregulated in 2 or more models") 
  ggsave("output/images/defect/enrichment/GO/GOBP_updefectgeq2.pdf",
         width = 7,
         height = 2.5)
pathway_enrichment_barplot(GO_BP_overrep_defect_geq2[[3]],
                           num_pathways = "sig_only",
                           sig_line = TRUE,
                           sig_cutoff = 0.05,
                           colour = "blue")+ 
  ggtitle("Biological processes with dephosphorylation dysregulated in 2 or more models") 
  ggsave("output/images/defect/enrichment/GO/GOBP_downdefectgeq2.pdf",
         width = 10,
         height = 3)

#MF
pathway_enrichment_barplot(GO_MF_overrep_defect_geq2[[1]],
                           num_pathways = "sig_only",
                           sig_line = TRUE,
                           sig_cutoff = 0.05,
                           colour = "grey") + 
  ggtitle("Molecular functions with dysregulation in 2 or more models") 
  ggsave("output/images/defect/enrichment/GO/GOMF_defectgeq2.pdf",
         width = 7,
         height = 2.5)
pathway_enrichment_barplot(GO_MF_overrep_defect_geq2[[2]],
                           num_pathways = "sig_only",
                           sig_line = TRUE,
                           sig_cutoff = 0.05,
                           colour = "red")+ 
  ggtitle("Molecular functions with phosphorylation dysregulated in 2 or more models") 
  ggsave("output/images/defect/enrichment/GO/GOMF_updefectgeq2.pdf",
         width = 7,
         height = 2.5)
pathway_enrichment_barplot(GO_MF_overrep_defect_geq2[[3]],
                           num_pathways = "sig_only",
                           sig_line = TRUE,
                           sig_cutoff = 0.05,
                           colour = "blue")+ 
  ggtitle("Molecular functions with dephosphorylation dysregulated in 2 or more models") 
  ggsave("output/images/defect/enrichment/GO/GOMF_downdefectgeq2.pdf",
         width = 7,
         height = 2.5)

#CC
pathway_enrichment_barplot(GO_CC_overrep_defect_geq2[[1]],
                           num_pathways = "sig_only",
                           sig_line = TRUE,
                           sig_cutoff = 0.05,
                           colour = "grey") + 
  ggtitle("Cellular compartments with dysregulation in 2 or more models") 
  ggsave("output/images/defect/enrichment/GO/GOCC_defectgeq2.pdf",
         width = 7,
         height = 2.5)
pathway_enrichment_barplot(GO_CC_overrep_defect_geq2[[2]],
                           num_pathways = "sig_only",
                           sig_line = TRUE,
                           sig_cutoff = 0.05,
                           colour = "red")+ 
  ggtitle("Cellular compartments with phosphorylation dysregulated in 2 or more models") 
  ggsave("output/images/defect/enrichment/GO/GOCC_updefectgeq2.pdf",
         width = 7,
         height = 2.5)
pathway_enrichment_barplot(GO_CC_overrep_defect_geq2[[3]],
                           num_pathways = "sig_only",
                           sig_line = TRUE,
                           sig_cutoff = 0.05,
                           colour = "blue")+ 
  ggtitle("Cellular compartments with dephosphorylation dysregulated in 2 or more models") 
  ggsave("output/images/defect/enrichment/GO/GOCC_downdefectgeq2.pdf",
         width = 7,
         height = 2.5)


###Fused

GO_CC_overrep_defect_geq2_updown_plot <- 
  pathway_enrichment_barplot_fuse(list(GO_CC_overrep_defect_geq2[[2]],
                                     GO_CC_overrep_defect_geq2[[3]]),
                                 num_pathways = "sig_only",
                           sig_line = TRUE,
                           sig_cutoff = 0.05,
                           colours = c("red",
                                       "blue"))

ggsave(GO_CC_overrep_defect_geq2_updown_plot,
       file = "output/images/defect/enrichment/GO/GOCC_up_and_downdefectgeq2.pdf",
       height = 6,
       width = 9)

GO_BP_overrep_defect_geq2_updown_plot <- 
  pathway_enrichment_barplot_fuse(list(GO_BP_overrep_defect_geq2[[2]],
                                     GO_BP_overrep_defect_geq2[[3]]),
                                 num_pathways = "sig_only",
                           sig_line = TRUE,
                           sig_cutoff = 0.05,
                           colours = c("red",
                                       "blue"))
ggsave(GO_BP_overrep_defect_geq2_updown_plot,
       file = "output/images/defect/enrichment/GO/GOBP_up_and_downdefectgeq2.pdf",
       height = 6,
       width = 11)

###Export
GO_output_dir <- "output/data/defect/enrichment/defect_geq2_GO_"
write.csv(GO_BP_overrep_defect_geq2$up,
          file = paste(GO_output_dir, "BP_up", ".csv", sep = ""))
write.csv(GO_BP_overrep_defect_geq2$down,
          file = paste(GO_output_dir, "BP_down", ".csv", sep = ""))
write.csv(GO_CC_overrep_defect_geq2$up,
          file = paste(GO_output_dir, "CC_up", ".csv", sep = ""))
write.csv(GO_CC_overrep_defect_geq2$down,
          file = paste(GO_output_dir, "CC_down", ".csv", sep = ""))
```


####BAS vs INS defects
Test for BAS and INS defects
```{r}
source("..\\scripts/3T3L1_BASorINS_defect_fns.R")
###Function that tests for BAS and INS defects, makes summary plot, and exports data
#num_data: THe data that will be tested e.g. med_norm data, data normalised to protein
BASorINSdef_master <- function(data, num_data,
                               image_file, data_file,
                               ...){
  data_BASorINSdefstats <- defective_BASorINS_bothdf(data, num_data)
  data_BASorINSdef <-  detect_BASorINS_defects(data_BASorINSdefstats,
                                               ...)
  #Plot and export
  BASorINSdefect_summary_bplot(data_BASorINSdef, ...)
  ggsave_pdfpng(image_file,
                  width = 2.5, height = 2.5)
  save(data_BASorINSdef,
       file = paste(data_file, ".RData", sep = ""))
  return(data_BASorINSdef)
}


#mednorm, 0.58 cutoff

phos_3t3_proc_mednormBASorINSdef <-  BASorINSdef_master(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef, 
                                                        phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef[, 1:52],
                   image_file = "output/images/defect/summary/BASorINS_defect/bplot_BASorINSdef_mednorm0.58",
                   data_file = "output/data/defect/BASorINS_defect/phos_3t3_proc_mednormBASorINSdef__1")


##Separating up and down sites

#Pattern bplot
BASorINSdefect_summary_bplot(phos_3t3_proc_mednormBASorINSdef,
                             summary = BASorINSdefect_summariser_updownsep(phos_3t3_proc_mednormBASorINSdef))
ggsave_pdfpng("output/images/defect/summary/BASorINS_defect/bplot_BASorINSdefupdown_mednorm0.58",
              width = 4, height = 2.5)
```



###Export
```{r}
save(phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef,
     file = "output/data/phos_3t3_filt_norm_imp_below_ds1_looserstats_stats_cols_regdef__1.RData")
```
