==> 01_3T3L1_phos/09__core_ins_sig.Rmd <==
---
title: "12__core_ins_sig__3"
output: html_document
---


Background

Here I look at canonical insulin signalling proteins and their sites


###Initialise
```{r}
library(reshape2)
library(ggplot2)
library(gridExtra)
source("..\\scripts/essential_visualisations_JvG.R")
source("..\\scripts/boxplots_JvG.R")
source("..\\scripts/3T3L1_mouse_IR__DF__202104_objects.R")
source("..\\scripts/3T3L1_mouse_IR__DF__202104_functions.R")
```

###Load in data
```{r}
load("data/intermediate/phos_3t3/phos_3t3_proc__4.RData")
```


###Canonical insulin signalling network

Read in, process, all sites and reg sites
```{r}
###Read in and process data

##Read in
can_ins_sig_proteins_EN <- readRDS("data/intermediate/can_ins_sig/ins_and_sean.rds")

##Decaps
can_ins_sig_proteins_EN <- sapply(can_ins_sig_proteins_EN,
                                  function(name){
                                    
                                    lower_name <- tolower(name)
                                    name_list <- strsplit(lower_name,
                                                          "")[[1]]
                                    name_list[1] <- toupper(name_list[1])
                                    final_name <- paste(name_list,
                                                        collapse = "")
                                  })

##All sites in 3t3l1
can_ins_sig_3t3l1_sites_EN <- rownames(phos_3t3_proc[which(phos_3t3_proc$gene %in% 
                                                                            can_ins_sig_proteins_EN), ])

##All sites regulated in 3t3l1
can_ins_sig_3t3l1_reg_sites_EN <- can_ins_sig_3t3l1_sites_EN[which(can_ins_sig_3t3l1_sites_EN %in%
                                                                     rownames(phos_3t3_proc[phos_3t3_proc$CTRL_reg != "unregulated", ]))]

```

###Defect sites
```{r}
####Summaries and plots

###Old defect filtering

##Summary
colnames(phos_3t3_proc) <- gsub("old",
                                                                        "",
                                                                        colnames(phos_3t3_proc))
can_ins_sig_olddefect_summary <- 
  reg_sites_summariser(phos_3t3_proc[can_ins_sig_3t3l1_sites_EN, ],
                       all_def_row = TRUE)
can_ins_sig_olddefect_summary_perc <- 
  reg_sites_summariser(phos_3t3_proc[can_ins_sig_3t3l1_sites_EN, ],
                       perc = TRUE,
                       all_def_row = TRUE)
reg_olddefect_summary_perc <- reg_sites_summariser(phos_3t3_proc,
                                                   perc = TRUE,
                                                   all_def_row = TRUE)


#Set up df
can_ins_sig_olddefect_comb_summary_perc <- cbind(reg_olddefect_summary_perc[, 1:4],
                                            can_ins_sig_olddefect_summary_perc[, 2:4])

colnames(can_ins_sig_olddefect_comb_summary_perc)  <- c("num_defective_models",
                                                   "reg_all_models_perc",
                                                            "reg_up_models_perc",
                                                            "reg_down_models_perc",
                                                            "can_all_models_perc",
                                                            "can_up_models_perc",
                                                            "can_down_models_perc")


###Dodged bar plot comparing perc of defect sites sites to all ins reg
can_ins_sig_olddefect_comb_summary_perc_melt <- melt(can_ins_sig_olddefect_comb_summary_perc[, c("num_defective_models",
    "can_all_models_perc",
    "reg_all_models_perc")],
    id.vars = "num_defective_models")
can_ins_sig_olddefect_comb_summary_perc_melt$num_defective_models <- 
  factor(can_ins_sig_olddefect_comb_summary_perc_melt$num_defective_models,
                                                             levels = c("0",
                                                                        "all_defective_sites",
                                                                        "1",
                                                                        "2",
                                                                        "3",
                                                                        "4",
                                                                        "5"))

ggplot(data = can_ins_sig_olddefect_comb_summary_perc_melt,
       aes(x = num_defective_models,
           y = value,
           fill = variable)) +
  geom_col(position = position_dodge()) +
  scale_fill_manual(values = c("#414141",
                               "#848484"),
                    labels = c("Regulated on canonical proteins",
                               "All regulated"),
                    name = "Site group") +
    ggtitle("Insulin-regulated peptides overall and on canonical proteins") +
    theme(panel.grid = element_blank(),
        panel.background = element_blank(),
        axis.line = element_line(color = "black"),
        plot.title = element_text(hjust = 0.5,
                                  size = 8),
        axis.text = element_text(colour = "black",
                                 size = 8),
        axis.title = element_text(colour = "black",
                                  size = 8),
        legend.text = element_text(colour = "black",
                                   size = 8),
        legend.title = element_text(colour = "black",
                                    size = 8),
        legend.key.size = unit(0.1,
                                   units = "inch")) +
  labs(x = "Number of defective treatments",
       y = "% of site group",
       size = 14) 
  ggsave("output/images/core_ins_sig/perc_barplot_olddefect_core_vs_bg.pdf",
         width = 4,
         height = 3)
```

Fisher's exact for lack of defect
```{r}
###Set up
reg_sites <- rownames(phos_3t3_proc[which(phos_3t3_proc$CTRL_reg != "unregulated"), ])
def_sites <- rownames(phos_3t3_proc[which(phos_3t3_proc$is_defective == TRUE), ])

k <- length(intersect(can_ins_sig_3t3l1_reg_sites_EN,
                                                def_sites))
K <- length(can_ins_sig_3t3l1_reg_sites_EN)
n <- length(def_sites)
N <- length(reg_sites)

contingency_table <- matrix(c(k,
                              K - k,
                              n - k,
                              N - n - K + k),
                            ncol = 2,
                            nrow = 2)

###Test
fisher.test(x = contingency_table,
            alternative = "less")


###Likelihoods of defect on canonical vs noncanonical
#Can: 0.2466216
#Not can: 0.4193353
length(intersect(def_sites,
                 can_ins_sig_3t3l1_reg_sites_EN)) / length(can_ins_sig_3t3l1_reg_sites_EN)

length(setdiff(def_sites,
               can_ins_sig_3t3l1_reg_sites_EN)) / length(setdiff(reg_sites,
                                                                 can_ins_sig_3t3l1_reg_sites_EN))
```


Sites with defect across multiple models
```{r}
phos_3t3_proc[which(rownames(phos_3t3_proc) %in% 
                       can_ins_sig_3t3l1_reg_sites_EN &
                      phos_3t3_proc$num_defective_models >= 3), ]

##2 sites on canonical proteins that are defective in >= 1 models
ggsave(boxplot_all_conditions_3t3_post202104_mult_panel(data = phos_3t3_proc,
                                                 sites = c("Tsc1_Q9EP53_S502_1",
                                "Deptor_Q570Y9_S244_1"),
                                ncol = 2,
                                nrow = 1),
       file = "output/images/core_ins_sig/core_reg_sites_defgeq1_2examples.pdf",
       width = 4,
       height = 1.5)

#Simpler boxplot
boxplot_3t3phos(phos_3t3_proc, c("Tsc1_Q9EP53_S502_1",
                                "Deptor_Q570Y9_S244_1"),
                facet = TRUE,
                facet_labels = c("Tsc1_Q9EP53_S502_1",
                                "Deptor_Q570Y9_S244_1"),
                facet_ncol_nrow = c(1, 2),
                siglabels = TRUE)
ggsave_pdfpng("output/images/core_ins_sig/core_reg_sites_defgeq1_2examples_simpler.pdf",
       width = 2,
       height = 3)
```

###ERK1/2
Here I inspect the activation sites on ERK1/2
```{r}
#Boxplots
ERK12_reg_sites <- c("Mapk1_P63085_T183_2",
                     "Mapk1_P63085_Y185_1",
                     "Mapk3_Q63844_T203_2",
                     "Mapk3_Q63844_Y205_1")

boxplot_3t3phos(phos_3t3_proc,
                site = ERK12_reg_sites,
                facet = TRUE,
                facet_labels = sapply(ERK12_reg_sites, function(x){
                  x_split <- strsplit(x, "_")[[1]]
                  paste(c(x_split[1], " ", x_split[3], " P", x_split[4]), collapse = "")
                }),
                facet_ncol_nrow = c(4, 1))
ggsave_pdfpng("output/images/core_ins_sig/canonical_sites/ERK12_reg_sites",
              height = 1.8, width = 7)


###Anovas and dunnett's on regulatort sites of ERK12
ERK12_reg_sites_dunnetts <- map(ERK12_reg_sites,
                                ~site_level_anova_dunnetts_3t3l1(phos_3t3_proc[, 1:52], .))
names(ERK12_reg_sites_dunnetts) <- ERK12_reg_sites

ERK12_reg_sites_dunnetts_df <- map(ERK12_reg_sites_dunnetts, 
                                   ~c(.$BAS_dunnetts_ps, .$INS_dunnetts_ps)) %>%
  purrr::reduce(rbind) %>% as.data.frame
rownames(ERK12_reg_sites_dunnetts_df) <- names(ERK12_reg_sites_dunnetts)
colnames(ERK12_reg_sites_dunnetts_df) <- cross(list(c("CI", "DEX", "TNF", "MPQ", "AA"), 
                                          c("bas", "ins"))) %>%
  map(~paste(unlist(.), collapse = "_")) %>% unlist
colnames(ERK12_reg_sites_dunnetts_df) <- paste(colnames(ERK12_reg_sites_dunnetts_df),
                                               "_p", sep = "")

write.csv(ERK12_reg_sites_dunnetts_df,
          file = "output/data/can_ins/ERK12_reg_sites_dunnetts_df.csv")
```



###Big heatmap
```{r}
###Heatmaps

# keep if all FCs, ordered by NORMAL_FC, squashed
fc_sig_heatmap(data = phos_3t3_proc[which(rownames(phos_3t3_proc) 
                                                                                           %in% can_ins_sig_3t3l1_reg_sites_EN &
                                                                          rowSums(is.na(phos_3t3_proc[, c("CTRL_FC",
                                                                                                                "CI_FC",
                                                                                                                "DEX_FC",
                                                                                                                "TNF_FC",
                                                                                                                "MPQ_FC",
                                                                                                                "AA_FC")])) == 0), ],
               fc_cols_wo_pvals = c("CTRL_FC",
                           "CI_FC",
                           "DEX_FC",
                           "TNF_FC",
                           "MPQ_FC",
                           "AA_FC"),
               order_column = "CTRL_FC",
               is_decreasing = FALSE,
               x_axis_names = c("NORMAL",
                                "CI",
                                "DEX",
                                "TNF",
                                "MPQ",
                                "AA"),
               colour_scheme = "rev_RdBu",
               tile_border_size = 0.05) + 
  coord_fixed(ratio = 0.2) +
  ggtitle("Regulated sites on core proteins") 
       ggsave("output/images/core_ins_sig/core_ins_sig_reg_allquant_nodot_brewer.pdf",
              width = 12,
              height = 20)
       ggsave("output/images/core_ins_sig/core_ins_sig_reg_allquant_nodot_brewer.png",
              width = 12,
              height = 20)
```


####Canonical insulin signalling sites
```{r}
###One big plot of notable canonical sites
can_ins_sig_reg_sites_toplot <- c("Akt1s1_Q9D1F4_T247_1",
                       "Rps6_P62754_S235_3",
                       "Rps6_P62754_S236_2",
                       "Mtor_Q9JLN9_S2481_2",
                       "Mapk1_P63085_Y185_1",
                       "Mapk3_Q63844_T203_2",
                       "Mapk3_Q63844_Y205_1",
                       "Tbc1d4_Q8BYJ6_S595_1",
                       "Gys1_Q9Z1E4_S641_2")

#Simpler boxplots
boxplot_3t3phos(data = phos_3t3_proc,
                site = can_ins_sig_reg_sites_toplot,
                facet = TRUE,
                facet_ncol_nrow = c(3, 3),
                facet_labels = sapply(can_ins_sig_reg_sites_toplot, function(x){
                  x_split <- strsplit(x, "_")[[1]]
                  return(paste(c(x_split[1], " ", x_split[3], " P", x_split[4]), collapse = ""))
                }),
                siglabels = TRUE,
                is_raw = TRUE) 
ggsave_pdfpng(file = "output/images/core_ins_sig/canonical_sites/supp_sites/canonical_ins_sig_supp_sites_noimplab_simpler.pdf",
       width = 6,
       height = 4.5)
```
