==> 01_3T3L1_phos/10__emergent.Rmd <==
---
title: "13__emergent__1"
output: html_document
---

Background

Here I look at emergent sites



Version info


###Initialise
```{r}
library(reshape2)
library(ggplot2)
library(tidyverse)
source("..\\scripts/essential_visualisations_JvG.R")
source("..\\scripts/enrichment_analysis_JvG.R")
source("..\\scripts/boxplots_JvG.R")
source("..\\scripts/3T3L1_mouse_IR__DF__202104_objects.R")
source("..\\scripts/3T3L1_mouse_IR__DF__202104_functions.R")
```

###Load in data
```{r}
load("data/intermediate/phos_3t3/phos_3t3_proc__4.RData")
```

###Filtering
```{r}
###Rules
##Control: Non-sig, |FC| < 0.58
##Sig in ANVOA
##Drugs: Sig in posthoc, |FC| > 0.58, agree

##Candidates
phos_3t3_proc$CTRL_no_reg <- FALSE
phos_3t3_proc$CTRL_no_reg[-which(phos_3t3_proc$t_adj_p_val < 0.05 |
        abs(phos_3t3_proc$CTRL_FC) > 0.58)] <- TRUE
##Emergent for each drug
phos_3t3_proc$MPQ_emergent <- FALSE
phos_3t3_proc$MPQ_emergent[which(phos_3t3_proc$CTRL_no_reg == TRUE &
                                                phos_3t3_proc$normal_anova_p_adj < 0.05 &
                                                 phos_3t3_proc$MPQ_padj < 0.05 &
                                                  phos_3t3_proc$MPQ_FC > 0.58)] <- "up"
phos_3t3_proc$MPQ_emergent[which(phos_3t3_proc$CTRL_no_reg == TRUE &
                                                phos_3t3_proc$normal_anova_p_adj < 0.05 &
                                                 phos_3t3_proc$MPQ_padj < 0.05 &
                                                  phos_3t3_proc$MPQ_FC < -0.58)] <- "down"
phos_3t3_proc$AA_emergent <- FALSE
phos_3t3_proc$AA_emergent[which(phos_3t3_proc$CTRL_no_reg == TRUE &
                                                phos_3t3_proc$normal_anova_p_adj < 0.05 &
                                                 phos_3t3_proc$AA_padj < 0.05 &
                                                  phos_3t3_proc$AA_FC > 0.58)] <- "up"
phos_3t3_proc$AA_emergent[which(phos_3t3_proc$CTRL_no_reg == TRUE &
                                                phos_3t3_proc$normal_anova_p_adj < 0.05 &
                                                 phos_3t3_proc$AA_padj < 0.05 &
                                                  phos_3t3_proc$AA_FC < -0.58)] <- "down"
phos_3t3_proc$CI_emergent <- FALSE
phos_3t3_proc$CI_emergent[which(phos_3t3_proc$CTRL_no_reg == TRUE &
                                                phos_3t3_proc$normal_anova_p_adj < 0.05 &
                                                 phos_3t3_proc$CI_padj < 0.05 &
                                                  phos_3t3_proc$CI_FC > 0.58)] <- "up"
phos_3t3_proc$CI_emergent[which(phos_3t3_proc$CTRL_no_reg == TRUE &
                                                phos_3t3_proc$normal_anova_p_adj < 0.05 &
                                                 phos_3t3_proc$CI_padj < 0.05 &
                                                  phos_3t3_proc$CI_FC < -0.58)] <- "down"
phos_3t3_proc$DEX_emergent <- FALSE
phos_3t3_proc$DEX_emergent[which(phos_3t3_proc$CTRL_no_reg == TRUE &
                                                phos_3t3_proc$normal_anova_p_adj < 0.05 &
                                                 phos_3t3_proc$DEX_padj < 0.05 &
                                                  phos_3t3_proc$DEX_FC > 0.58)] <- "up"
phos_3t3_proc$DEX_emergent[which(phos_3t3_proc$CTRL_no_reg == TRUE &
                                                phos_3t3_proc$normal_anova_p_adj < 0.05 &
                                                 phos_3t3_proc$DEX_padj < 0.05 &
                                                  phos_3t3_proc$DEX_FC < -0.58)] <- "down"
phos_3t3_proc$TNF_emergent <- FALSE
phos_3t3_proc$TNF_emergent[which(phos_3t3_proc$CTRL_no_reg == TRUE &
                                                phos_3t3_proc$normal_anova_p_adj < 0.05 &
                                                 phos_3t3_proc$TNF_padj < 0.05 &
                                                  phos_3t3_proc$TNF_FC > 0.58)] <- "up"
phos_3t3_proc$TNF_emergent[which(phos_3t3_proc$CTRL_no_reg == TRUE &
                                                phos_3t3_proc$normal_anova_p_adj < 0.05 &
                                                 phos_3t3_proc$TNF_padj < 0.05 &
                                                  phos_3t3_proc$TNF_FC < -0.58)] <- "down"

##Synthesising into emergent sites
phos_3t3_proc$emergent_up_num <- rowSums(phos_3t3_proc[, grep("_emergent",
                                    colnames(phos_3t3_proc))] == "up",
                                    na.rm = TRUE)
phos_3t3_proc$emergent_down_num <- rowSums(phos_3t3_proc[, grep("_emergent",
                                    colnames(phos_3t3_proc))] == "down",
                                    na.rm = TRUE)
#Histograms
hist(phos_3t3_proc$emergent_up_num[which(phos_3t3_proc$emergent_up_num > 0)],
     main = "up")
hist(phos_3t3_proc$emergent_down_num[which(phos_3t3_proc$emergent_down_num > 0)],
     main = "down")

#Sites that have up and down
phos_3t3_proc$emergent_up_and_down <- FALSE
phos_3t3_proc$emergent_up_and_down[which(phos_3t3_proc$emergent_up_num > 0 &
        phos_3t3_proc$emergent_down_num > 0)] <- TRUE

#Finally filter for emergent sites: Must be emergent in at least one model, and not up and down
phos_3t3_proc$emergent <- FALSE
phos_3t3_proc$emergent[which((phos_3t3_proc$emergent_up_num > 0 |
                                                            phos_3t3_proc$emergent_down_num > 0) &
                                                              phos_3t3_proc$emergent_up_and_down == FALSE)] <- TRUE

#Number of emergent models:
phos_3t3_proc$emergent_num <- 0
phos_3t3_proc$emergent_num[which(phos_3t3_proc$emergent == TRUE)] <- phos_3t3_proc$emergent_up_num[which(phos_3t3_proc$emergent == TRUE)] + 
  phos_3t3_proc$emergent_down_num[which(phos_3t3_proc$emergent == TRUE)]

##Summary
#896 emergent sites
#112 emergent in >= 2 models
#19 emergent in >= 3 models
#3 emergent in >= 4 models
#0 emergent in 5 models

length(which(phos_3t3_proc$emergent == TRUE))
length(which(phos_3t3_proc$emergent_num >= 2))
length(which(phos_3t3_proc$emergent_num >= 3))
length(which(phos_3t3_proc$emergent_num >= 4))
length(which(phos_3t3_proc$emergent_num >= 5))


emergent_summary <- emergent_sites_summariser(phos_3t3_proc)
write.csv(emergent_summary,
          file = "output/data/emergent/emergent_summary_3T3l1_DIA.csv")
```



###Summary
```{r}
emerg_summary <- emergent_sites_summariser(phos_3t3_proc) %>%
  rbind(c(0, colSums(.[, -1])))
emerg_summary[6, 1] <- "all_emergent"

###Bargraphs
library(tidyverse)

##Up and down
emerg_summary_up_vs_down <- pivot_longer(emerg_summary[, c("num_emergent_models",
                                                           "all_models_up",
                                                           "all_models_down")],
                                         cols = c("all_models_up",
                                                  "all_models_down"),
                                         names_to = "variable", values_to = "value") %>%
  mutate(num_emergent_models = factor(num_emergent_models,
                                      levels = c("all_emergent", 1:5))) %>%
  mutate(variable = factor(variable, levels = c("all_models_up", "all_models_down")))
levels(emerg_summary_up_vs_down$variable) <- c("up", "down")

ggplot(emerg_summary_up_vs_down,
       aes(x = num_emergent_models,
           y = value,
           fill = variable)) + 
  geom_col(position = "stack",
           width = 0.8) +
  scale_fill_manual(name = "Insulin Regulation",
                      labels = c("Upregulated",
                                 "Downregulated"),
                      values = up_down_colours) +
  comfy_theme(include_legend = FALSE) +
  theme(axis.line.x = element_blank(),
        axis.ticks.x = element_blank()) +
  labs(x = "Number of emergent models",
       y = "Emergent phosphopeptides") +
  ggtitle("Emergent phosphopeptides") 
  ggsave_pdfpng("output/images/emergent/summary/3t3l1_emerg_sites_up_down",
         width = 1.5,
         height = 2)
  
  
#Each model
  
##Num sig_fc for each drug
emerg_summary_models <- pivot_longer(emerg_summary[, c("num_emergent_models",
                                                       "CI_total",
                                                               "DEX_total",
                                                               "TNF_total",
                                                               "MPQ_total",
                                                               "AA_total")],
                                     cols = c("CI_total",
                                                               "DEX_total",
                                                               "TNF_total",
                                                               "MPQ_total",
                                                               "AA_total"),
                                     names_to = "variable", values_to = "value") %>%
  mutate(num_emergent_models = as.factor(num_emergent_models),
         model = factor(gsub("_total", "", variable), levels = c("CI",
                                                               "DEX",
                                                               "TNF",
                                                               "MPQ",
                                                               "AA")))

#x-axis model, alpha number of defective models
ggplot(subset(emerg_summary_models,
              num_emergent_models != "all_emergent"),
       aes(x = model,
           y = value,
           fill = model,
           alpha = num_emergent_models)) +
  geom_col(position = "stack",
           width = 0.8) +
  scale_fill_manual(values = SH_model_main_colours,
                    name = "Model",
                    labels = c("CI",
                               "DEX",
                               "TNF",
                               "MPQ",
                               "AA")) +
  scale_alpha_manual(values = rev(seq(from = 0.2,
                                  to = 1,
                                  by = (1 - 0.2)/4)),
                     name = "# emergent models")  +
  labs(x = "Model",
       y = "Emergent phosphopeptides") +
 comfy_theme(include_legend = FALSE) +
  theme(axis.line.x = element_blank(),
        axis.ticks.x = element_blank(),
        axis.text.x = element_text(colour = "black", size = 7, hjust = 1, vjust = 0.5, angle = 90)) +
  ggtitle("Emergent phosphopeptides") 
  ggsave_pdfpng("output/images/emergent/summary/3t3l1_emerg_sites_xaxis_model",
         width = 1.35,
         height = 2)
```



###Visualise
```{r}
##Emergent in >= 3 models
  
#Without outlier
fc_sig_heatmap(data = phos_3t3_proc[which(phos_3t3_proc$emergent_num >= 3 &
                                            phos_3t3_proc$gene != "Ppp2r5b"), ],
               fc_cols_w_pvals = c("CTRL_FC",
                           "CI_FC",
                           "DEX_FC",
                           "TNF_FC",
                           "MPQ_FC",
                           "AA_FC"),
               pval_cols = c("t_adj_p_val",
                             "CI_padj",
                             "DEX_padj",
                             "TNF_padj",
                             "MPQ_padj",
                             "AA_padj"),
               order_column = "TNF_FC",
               is_decreasing = FALSE,
               x_axis_names = c("CTRL",
                                "CI",
                                "DEX",
                                "TNF",
                                "MPQ",
                                "AA"),
               scale_colours_manually = TRUE) +
  ggtitle("Emergent in 3 or more models") 
  ggsave("output/images/emergent/emergent_geq3/emergent_geq3_DIA_noout.pdf",
         width = 4,
         height = 5)
  
#Simpler boxplot for outlier
boxplot_3t3phos(phos_3t3_proc,
                "Ppp2r5b_Q6PD28_S13_1",
                siglabels = TRUE,
                is_raw = TRUE)
ggsave_pdfpng("output/images/emergent/emergent_geq3/outlier_Ppp2r5b_Q6PD28_S13_1",
              width = 2, height = 1.8)
```


####Enrichment

###GO
```{r}
###Emergent geq 3

##BP
GO_BP_emergent_geq3 <- list()
GO_BP_emergent_geq3[[1]] <- GO_enricher(background_genes = unique(phos_3t3_proc$gene),
                        DE_genes = unique(phos_3t3_proc$gene[which(phos_3t3_proc$emergent_num >= 3)]),
                        gene_label = "SYMBOL",
                        organism_database = org.Mm.eg.db,
                        ontology = "BP",
                        pathway_DE_intersection_threshold = 3)
GO_BP_emergent_geq3[[2]] <- GO_enricher(background_genes = unique(phos_3t3_proc$gene),
                        DE_genes = unique(phos_3t3_proc$gene[which(phos_3t3_proc$emergent_num >= 3 &
                                                                     phos_3t3_proc$emergent_up_num >= 1)]),
                        gene_label = "SYMBOL",
                        organism_database = org.Mm.eg.db,
                        ontology = "BP",
                        pathway_DE_intersection_threshold = 3)
GO_BP_emergent_geq3[[3]] <- GO_enricher(background_genes = unique(phos_3t3_proc$gene),
                        DE_genes = unique(phos_3t3_proc$gene[which(phos_3t3_proc$emergent_num >= 3 &
                                                                     phos_3t3_proc$emergent_down_num >= 1)]),
                        gene_label = "SYMBOL",
                        organism_database = org.Mm.eg.db,
                        ontology = "BP",
                        pathway_DE_intersection_threshold = 3)
names(GO_BP_emergent_geq3) <- c("up_or_down",
                                "up",
                                "down")

##MF
GO_MF_emergent_geq3 <- list()
GO_MF_emergent_geq3[[1]] <- GO_enricher(background_genes = unique(phos_3t3_proc$gene),
                        DE_genes = unique(phos_3t3_proc$gene[which(phos_3t3_proc$emergent_num >= 3)]),
                        gene_label = "SYMBOL",
                        organism_database = org.Mm.eg.db,
                        ontology = "MF",
                        pathway_DE_intersection_threshold = 3)
GO_MF_emergent_geq3[[2]] <- GO_enricher(background_genes = unique(phos_3t3_proc$gene),
                        DE_genes = unique(phos_3t3_proc$gene[which(phos_3t3_proc$emergent_num >= 3 &
                                                                     phos_3t3_proc$emergent_up_num >= 1)]),
                        gene_label = "SYMBOL",
                        organism_database = org.Mm.eg.db,
                        ontology = "MF",
                        pathway_DE_intersection_threshold = 3)
GO_MF_emergent_geq3[[3]] <- GO_enricher(background_genes = unique(phos_3t3_proc$gene),
                        DE_genes = unique(phos_3t3_proc$gene[which(phos_3t3_proc$emergent_num >= 3 &
                                                                     phos_3t3_proc$emergent_down_num >= 1)]),
                        gene_label = "SYMBOL",
                        organism_database = org.Mm.eg.db,
                        ontology = "MF",
                        pathway_DE_intersection_threshold = 3)
names(GO_MF_emergent_geq3) <- c("up_or_down",
                                "up",
                                "down")

##CC
GO_CC_emergent_geq3 <- list()
GO_CC_emergent_geq3[[1]] <- GO_enricher(background_genes = unique(phos_3t3_proc$gene),
                        DE_genes = unique(phos_3t3_proc$gene[which(phos_3t3_proc$emergent_num >= 3)]),
                        gene_label = "SYMBOL",
                        organism_database = org.Mm.eg.db,
                        ontology = "CC",
                        pathway_DE_intersection_threshold = 3)
GO_CC_emergent_geq3[[2]] <- GO_enricher(background_genes = unique(phos_3t3_proc$gene),
                        DE_genes = unique(phos_3t3_proc$gene[which(phos_3t3_proc$emergent_num >= 3 &
                                                                     phos_3t3_proc$emergent_up_num >= 1)]),
                        gene_label = "SYMBOL",
                        organism_database = org.Mm.eg.db,
                        ontology = "CC",
                        pathway_DE_intersection_threshold = 3)
GO_CC_emergent_geq3[[3]] <- GO_enricher(background_genes = unique(phos_3t3_proc$gene),
                        DE_genes = unique(phos_3t3_proc$gene[which(phos_3t3_proc$emergent_num >= 3 &
                                                                     phos_3t3_proc$emergent_down_num >= 1)]),
                        gene_label = "SYMBOL",
                        organism_database = org.Mm.eg.db,
                        ontology = "CC",
                        pathway_DE_intersection_threshold = 3)
names(GO_CC_emergent_geq3) <- c("up_or_down",
                                "up",
                                "down")


###Visualise

#GO BP emergent up geq3
pathway_enrichment_barplot(GO_BP_emergent_geq3[[2]],
                           num_pathways = "sig_only",
                           sig_line = TRUE,
                           sig_cutoff = 0.1,
                           colour = "red") + 
  ggtitle("Biological processes emergent in 3 or more models") 
  ggsave("output/images/emergent/enrichment/GOBP_emergentupgeq3.pdf",
         width = 7,
         height = 3)

#Export
write.csv(GO_BP_emergent_geq3$up,
          file = "output/data/emergent/emergent_upgeq3_GOBPs.csv")
```
