==> 01_3T3L1_phos/11__proteome_overlap.Rmd <==
---
title: "14__proteome_overlap__1"
output: html_document
---


Background

Here I look at the overlap between changes in phospho and changes in proteome



Version info


###Initialise
```{r}
library(reshape2)
library(ggplot2)
library(gridExtra)
source("..\\scripts/3T3L1_mouse_IR__DF__202104_objects.R")
```

###Load in data
```{r}
load("data/intermediate/phos_3t3/phos_3t3_proc__4.RData")
```


###Defective global model vs ctrl changes


Big plots comparing proteome MODEL/CTRL and phospho BAS MODEL/CTRL or BAS MODEL/CTRL
```{r}
####BAS

###Find x and y limits
x_lim <- max(abs(c(phos_3t3_proc$CI_GROUPED_FC[which(phos_3t3_proc$CI_defect == TRUE)],
                   phos_3t3_proc$DEX_GROUPED_FC[which(phos_3t3_proc$DEX_defect == TRUE)],
                   phos_3t3_proc$TNF_GROUPED_FC[which(phos_3t3_proc$TNF_defect == TRUE)],
                   phos_3t3_proc$MPQ_GROUPED_FC[which(phos_3t3_proc$MPQ_defect == TRUE)],
                   phos_3t3_proc$AA_GROUPED_FC[which(phos_3t3_proc$AA_defect == TRUE)])),
             na.rm = TRUE)
y_lim <- max(abs(c(phos_3t3_proc$CI_FC_on_control_BAS[which(phos_3t3_proc$CI_defect == TRUE)],
                   phos_3t3_proc$DEX_FC_on_control_BAS[which(phos_3t3_proc$DEX_defect == TRUE)],
                   phos_3t3_proc$TNF_FC_on_control_BAS[which(phos_3t3_proc$TNF_defect == TRUE)],
                   phos_3t3_proc$MPQ_FC_on_control_BAS[which(phos_3t3_proc$MPQ_defect == TRUE)],
                   phos_3t3_proc$AA_FC_on_control_BAS[which(phos_3t3_proc$AA_defect == TRUE)])),
             na.rm = TRUE)

###Plot

models <- c("CI",
            "DEX",
            "TNF",
            "MPQ",
            "AA")
BAS_plots_list <- list()
for (i in 1:length(models)){
  
  ##Run correlation
  temp_test <- cor.test(phos_3t3_proc[which(phos_3t3_proc[, paste(c(models[i],
                                                                     "_defect"),
                                                                   collapse = "")] == TRUE), 
                                      paste(c(models[i],
                              "_GROUPED_FC"),
                            collapse = "")],
            phos_3t3_proc[which(phos_3t3_proc[, paste(c(models[i],
                                                                     "_defect"),
                                                                   collapse = "")] == TRUE), paste(c(models[i],
                       "_FC_on_control_BAS"),
                     collapse = "")],
            na.action("na.omit"))
  
  ##Make correlation data
  corr_plot_data <- data.frame("x" = 1,
                               "y_corr" = y_lim - 0.5,
                               "y_pval" = y_lim - 1.5,
                               "corr" = paste(c("r = ",
                                              round(temp_test$estimate,
                                                    digits = 3)),
                                              collapse = ""),
                               "pval" = paste(c("p = ",
                                              signif(temp_test$p.value,
                                                    digits = 3)),
                                              collapse = ""))
  
  ##Print corr data
  
  ##Plot and include corr
  temp_plot <- ggplot(phos_3t3_proc[which(phos_3t3_proc[, paste(c(models[i],
                                                                     "_defect"),
                                                                   collapse = "")] == TRUE), ],
       aes_string(x = paste(c(models[i],
                              "_GROUPED_FC"),
                            collapse = ""),
           y = paste(c(models[i],
                       "_FC_on_control_BAS"),
                     collapse = ""))) +
         geom_hline(yintercept = 0,
             colour = "black") +
  geom_vline(xintercept = 0,
             colour = "black") +
  geom_point(size = 1,
             shape = 21,
             stroke = 0.3,
             colour = alpha(SH_model_main_colours[i],
                            0.5),
             fill = alpha(SH_model_main_colours[i],
                          0.3)) +
    geom_text(data = corr_plot_data,
              aes(x = x,
                  y = y_corr,
                  label = corr),
              hjust = 0,
              size = 3) +
    geom_text(data = corr_plot_data,
              aes(x = x,
                  y = y_pval,
                  label = pval),
              hjust = 0,
              size = 3) +
  theme(panel.background = element_blank(),
        panel.grid = element_blank(),
        axis.text = element_text(colour = "black"),
        plot.title = element_text(hjust = 0.5)) +
    labs(x = "Proteome log2 MODEL/CTRL",
         y = "Phosphoproteome BAS log2 MODEL/CTRL") +
    xlim(-x_lim, x_lim) +
    ylim(-y_lim, y_lim) +
    ggtitle(models[i])
  
  
  #Add to list
  BAS_plots_list[[i]] <- temp_plot
  
  #plot
  plot(temp_plot)
  ggsave(paste(c("output/images/phospho_proteo_overlap/phos_prot_change_corr/defect_only/BAS_phos/",
                   models[i],
                   "_phos_prot_change_corr_defonly_BAS_phos_plot.pdf"),
                 collapse = ""),
           width = 2,
           height = 2)
    

}

#Save all combined
ggsave("output/images/phospho_proteo_overlap/phos_prot_change_corr/defect_only/BAS_phos/all_models_phos_prot_change_corr_defonly_BAS_phos_plot.pdf",
       arrangeGrob(grobs = BAS_plots_list,
                   ncol = 5),
       width = 10,
       height = 2)



####INS

###Find x and y limits
x_lim <- max(abs(c(phos_3t3_proc$CI_GROUPED_FC[which(phos_3t3_proc$CI_defect == TRUE)],
                   phos_3t3_proc$DEX_GROUPED_FC[which(phos_3t3_proc$DEX_defect == TRUE)],
                   phos_3t3_proc$TNF_GROUPED_FC[which(phos_3t3_proc$TNF_defect == TRUE)],
                   phos_3t3_proc$MPQ_GROUPED_FC[which(phos_3t3_proc$MPQ_defect == TRUE)],
                   phos_3t3_proc$AA_GROUPED_FC[which(phos_3t3_proc$AA_defect == TRUE)])),
             na.rm = TRUE)
y_lim <- max(abs(c(phos_3t3_proc$CI_FC_on_control_INS[which(phos_3t3_proc$CI_defect == TRUE)],
                   phos_3t3_proc$DEX_FC_on_control_INS[which(phos_3t3_proc$DEX_defect == TRUE)],
                   phos_3t3_proc$TNF_FC_on_control_INS[which(phos_3t3_proc$TNF_defect == TRUE)],
                   phos_3t3_proc$MPQ_FC_on_control_INS[which(phos_3t3_proc$MPQ_defect == TRUE)],
                   phos_3t3_proc$AA_FC_on_control_INS[which(phos_3t3_proc$AA_defect == TRUE)])),
             na.rm = TRUE)

###Plot

models <- c("CI",
            "DEX",
            "TNF",
            "MPQ",
            "AA")
INS_plots_list <- list()
for (i in 1:length(models)){
  
  ##Run correlation
  temp_test <- cor.test(phos_3t3_proc[which(phos_3t3_proc[, paste(c(models[i],
                                                                     "_defect"),
                                                                   collapse = "")] == TRUE), 
                                      paste(c(models[i],
                              "_GROUPED_FC"),
                            collapse = "")],
            phos_3t3_proc[which(phos_3t3_proc[, paste(c(models[i],
                                                                     "_defect"),
                                                                   collapse = "")] == TRUE), paste(c(models[i],
                       "_FC_on_control_INS"),
                     collapse = "")],
            na.action("na.omit"))
  
  ##Make correlation data
  corr_plot_data <- data.frame("x" = 1,
                               "y_corr" = y_lim - 0.5,
                               "y_pval" = y_lim - 1.5,
                               "corr" = paste(c("r = ",
                                              round(temp_test$estimate,
                                                    digits = 3)),
                                              collapse = ""),
                               "pval" = paste(c("p = ",
                                              signif(temp_test$p.value,
                                                    digits = 3)),
                                              collapse = ""))
  
  ##Print corr data
  
  ##Plot and include corr
  temp_plot <- ggplot(phos_3t3_proc[which(phos_3t3_proc[, paste(c(models[i],
                                                                     "_defect"),
                                                                   collapse = "")] == TRUE), ],
       aes_string(x = paste(c(models[i],
                              "_GROUPED_FC"),
                            collapse = ""),
           y = paste(c(models[i],
                       "_FC_on_control_INS"),
                     collapse = ""))) +
         geom_hline(yintercept = 0,
             colour = "black") +
  geom_vline(xintercept = 0,
             colour = "black") +
  geom_point(size = 1,
             shape = 21,
             stroke = 0.3,
             colour = alpha(SH_model_main_colours[i],
                            0.5),
             fill = alpha(SH_model_main_colours[i],
                          0.3)) +
    geom_text(data = corr_plot_data,
              aes(x = x,
                  y = y_corr,
                  label = corr),
              hjust = 0,
              size = 3) +
    geom_text(data = corr_plot_data,
              aes(x = x,
                  y = y_pval,
                  label = pval),
              hjust = 0,
              size = 3) +
  theme(panel.background = element_blank(),
        panel.grid = element_blank(),
        axis.text = element_text(colour = "black"),
        plot.title = element_text(hjust = 0.5)) +
    labs(x = "Proteome log2 MODEL/CTRL",
         y = "Phosphoproteome INS log2 MODEL/CTRL") +
    xlim(-x_lim, x_lim) +
    ylim(-y_lim, y_lim) +
    ggtitle(models[i])
  
  
  #Add to list
  INS_plots_list[[i]] <- temp_plot
  
  #plot
  plot(temp_plot)
  ggsave(paste(c("output/images/phospho_proteo_overlap/phos_prot_change_corr/defect_only/INS_phos/",
                   models[i],
                   "_phos_prot_change_corr_defonly_INS_phos_plot.pdf"),
                 collapse = ""),
           width = 2,
           height = 2)
    

}

#Save all combined
ggsave("output/images/phospho_proteo_overlap/phos_prot_change_corr/defect_only/INS_phos/all_models_phos_prot_change_corr_defonly_INS_phos_plot.pdf",
       arrangeGrob(grobs = INS_plots_list,
                   ncol = 5),
       width = 10,
       height = 2)
```


###Emergent global model vs ctrl changes
Big plots comparing proteome MODEL/CTRL and phospho BAS MODEL/CTRL or BAS MODEL/CTRL
```{r}
####BAS

###Find x and y limits
x_lim <- max(abs(c(phos_3t3_proc$CI_GROUPED_FC[which(phos_3t3_proc$CI_emergent != "FALSE")],
                   phos_3t3_proc$DEX_GROUPED_FC[which(phos_3t3_proc$DEX_emergent != "FALSE")],
                   phos_3t3_proc$TNF_GROUPED_FC[which(phos_3t3_proc$TNF_emergent != "FALSE")],
                   phos_3t3_proc$MPQ_GROUPED_FC[which(phos_3t3_proc$MPQ_emergent != "FALSE")],
                   phos_3t3_proc$AA_GROUPED_FC[which(phos_3t3_proc$AA_emergent != "FALSE")])),
             na.rm = TRUE)
y_lim <- max(abs(c(phos_3t3_proc$CI_FC_on_control_BAS[which(phos_3t3_proc$CI_emergent != "FALSE")],
                   phos_3t3_proc$DEX_FC_on_control_BAS[which(phos_3t3_proc$DEX_emergent != "FALSE")],
                   phos_3t3_proc$TNF_FC_on_control_BAS[which(phos_3t3_proc$TNF_emergent != "FALSE")],
                   phos_3t3_proc$MPQ_FC_on_control_BAS[which(phos_3t3_proc$MPQ_emergent != "FALSE")],
                   phos_3t3_proc$AA_FC_on_control_BAS[which(phos_3t3_proc$AA_emergent != "FALSE")])),
             na.rm = TRUE)

###Plot

models <- c("CI",
            "DEX",
            "TNF",
            "MPQ",
            "AA")
BAS_plots_list <- list()
for (i in 1:length(models)){
  
  ##Run correlation
  temp_test <- cor.test(phos_3t3_proc[which(phos_3t3_proc[, paste(c(models[i],
                                                                     "_emergent"),
                                                                   collapse = "")] != "FALSE"), 
                                      paste(c(models[i],
                              "_GROUPED_FC"),
                            collapse = "")],
            phos_3t3_proc[which(phos_3t3_proc[, paste(c(models[i],
                                                                     "_emergent"),
                                                                   collapse = "")] != "FALSE"), paste(c(models[i],
                       "_FC_on_control_BAS"),
                     collapse = "")],
            na.action("na.omit"))
  
  ##Make correlation data
  corr_plot_data <- data.frame("x" = 1,
                               "y_corr" = y_lim - 0.5,
                               "y_pval" = y_lim - 1.5,
                               "corr" = paste(c("r = ",
                                              round(temp_test$estimate,
                                                    digits = 3)),
                                              collapse = ""),
                               "pval" = paste(c("p = ",
                                              signif(temp_test$p.value,
                                                    digits = 3)),
                                              collapse = ""))
  
  ##Print corr data
  
  ##Plot and include corr
  temp_plot <- ggplot(phos_3t3_proc[which(phos_3t3_proc[, paste(c(models[i],
                                                                     "_emergent"),
                                                                   collapse = "")] != "FALSE"), ],
       aes_string(x = paste(c(models[i],
                              "_GROUPED_FC"),
                            collapse = ""),
           y = paste(c(models[i],
                       "_FC_on_control_BAS"),
                     collapse = ""))) +
         geom_hline(yintercept = 0,
             colour = "black") +
  geom_vline(xintercept = 0,
             colour = "black") +
  geom_point(size = 1,
             shape = 21,
             stroke = 0.3,
             colour = alpha(SH_model_main_colours[i],
                            0.5),
             fill = alpha(SH_model_main_colours[i],
                          0.3)) +
    geom_text(data = corr_plot_data,
              aes(x = x,
                  y = y_corr,
                  label = corr),
              hjust = 0,
              size = 3) +
    geom_text(data = corr_plot_data,
              aes(x = x,
                  y = y_pval,
                  label = pval),
              hjust = 0,
              size = 3) +
  theme(panel.background = element_blank(),
        panel.grid = element_blank(),
        axis.text = element_text(colour = "black"),
        plot.title = element_text(hjust = 0.5)) +
    labs(x = "Proteome log2 MODEL/CTRL",
         y = "Phosphoproteome BAS log2 MODEL/CTRL") +
    xlim(-x_lim, x_lim) +
    ylim(-y_lim, y_lim) +
    ggtitle(models[i])
  
  
  #Add to list
  BAS_plots_list[[i]] <- temp_plot
  
  #plot
  plot(temp_plot)
   ggsave(paste(c("output/images/phospho_proteo_overlap/phos_prot_change_corr/emergent_only/BAS_phos/",
                   models[i],
                   "_phos_prot_change_corr_emergentonly_BAS_phos_plot.pdf"),
                 collapse = ""),
           width = 2,
           height = 2)
    

}

#Save all combined
ggsave("output/images/phospho_proteo_overlap/phos_prot_change_corr/emergent_only/BAS_phos/allmod_phos_prot_emergentonly_BAS_phos_plot.pdf",
       arrangeGrob(grobs = BAS_plots_list,
                   ncol = 5),
       width = 10,
       height = 2)



####INS

###Find x and y limits
x_lim <- max(abs(c(phos_3t3_proc$CI_GROUPED_FC[which(phos_3t3_proc$CI_emergent != "FALSE")],
                   phos_3t3_proc$DEX_GROUPED_FC[which(phos_3t3_proc$DEX_emergent != "FALSE")],
                   phos_3t3_proc$TNF_GROUPED_FC[which(phos_3t3_proc$TNF_emergent != "FALSE")],
                   phos_3t3_proc$MPQ_GROUPED_FC[which(phos_3t3_proc$MPQ_emergent != "FALSE")],
                   phos_3t3_proc$AA_GROUPED_FC[which(phos_3t3_proc$AA_emergent != "FALSE")])),
             na.rm = TRUE)
y_lim <- max(abs(c(phos_3t3_proc$CI_FC_on_control_INS[which(phos_3t3_proc$CI_emergent != "FALSE")],
                   phos_3t3_proc$DEX_FC_on_control_INS[which(phos_3t3_proc$DEX_emergent != "FALSE")],
                   phos_3t3_proc$TNF_FC_on_control_INS[which(phos_3t3_proc$TNF_emergent != "FALSE")],
                   phos_3t3_proc$MPQ_FC_on_control_INS[which(phos_3t3_proc$MPQ_emergent != "FALSE")],
                   phos_3t3_proc$AA_FC_on_control_INS[which(phos_3t3_proc$AA_emergent != "FALSE")])),
             na.rm = TRUE)

###Plot

models <- c("CI",
            "DEX",
            "TNF",
            "MPQ",
            "AA")
INS_plots_list <- list()
for (i in 1:length(models)){
  
  ##Run correlation
  temp_test <- cor.test(phos_3t3_proc[which(phos_3t3_proc[, paste(c(models[i],
                                                                     "_emergent"),
                                                                   collapse = "")] != "FALSE"), 
                                      paste(c(models[i],
                              "_GROUPED_FC"),
                            collapse = "")],
            phos_3t3_proc[which(phos_3t3_proc[, paste(c(models[i],
                                                                     "_emergent"),
                                                                   collapse = "")] != "FALSE"), paste(c(models[i],
                       "_FC_on_control_INS"),
                     collapse = "")],
            na.action("na.omit"))
  
  ##Make correlation data
  corr_plot_data <- data.frame("x" = 1,
                               "y_corr" = y_lim - 0.5,
                               "y_pval" = y_lim - 1.5,
                               "corr" = paste(c("r = ",
                                              round(temp_test$estimate,
                                                    digits = 3)),
                                              collapse = ""),
                               "pval" = paste(c("p = ",
                                              signif(temp_test$p.value,
                                                    digits = 3)),
                                              collapse = ""))
  
  ##Print corr data
  
  ##Plot and include corr
  temp_plot <- ggplot(phos_3t3_proc[which(phos_3t3_proc[, paste(c(models[i],
                                                                     "_emergent"),
                                                                   collapse = "")] != "FALSE"), ],
       aes_string(x = paste(c(models[i],
                              "_GROUPED_FC"),
                            collapse = ""),
           y = paste(c(models[i],
                       "_FC_on_control_INS"),
                     collapse = ""))) +
         geom_hline(yintercept = 0,
             colour = "black") +
  geom_vline(xintercept = 0,
             colour = "black") +
  geom_point(size = 1,
             shape = 21,
             stroke = 0.3,
             colour = alpha(SH_model_main_colours[i],
                            0.5),
             fill = alpha(SH_model_main_colours[i],
                          0.3)) +
    geom_text(data = corr_plot_data,
              aes(x = x,
                  y = y_corr,
                  label = corr),
              hjust = 0,
              size = 3) +
    geom_text(data = corr_plot_data,
              aes(x = x,
                  y = y_pval,
                  label = pval),
              hjust = 0,
              size = 3) +
  theme(panel.background = element_blank(),
        panel.grid = element_blank(),
        axis.text = element_text(colour = "black"),
        plot.title = element_text(hjust = 0.5)) +
    labs(x = "Proteome log2 MODEL/CTRL",
         y = "Phosphoproteome INS log2 MODEL/CTRL") +
    xlim(-x_lim, x_lim) +
    ylim(-y_lim, y_lim) +
    ggtitle(models[i])
  
  
  #Add to list
  INS_plots_list[[i]] <- temp_plot
  
  #plot
  plot(temp_plot)
  ggsave(paste(c("output/images/phospho_proteo_overlap/phos_prot_change_corr/emergent_only/INS_phos/",
                   models[i],
                   "_phos_prot_change_corr_emergentonly_INS_phos_plot.pdf"),
                 collapse = ""),
           width = 2,
           height = 2)
    

}

#Save all combined
ggsave("output/images/phospho_proteo_overlap/phos_prot_change_corr/emergent_only/INS_phos/allmod_phos_prot_emergentonly_INS_phos_plot.pdf",
       arrangeGrob(grobs = INS_plots_list,
                   ncol = 5),
       width = 10,
       height = 2)
```
