==> 01_3T3L1_phos/12__GSK3.Rmd <==
---
title: "16__GSK3__3"
output: html_document
editor_options: 
  chunk_output_type: inline
---

Background

Here I map GSK3 inhibitor ppomics into 3T3 DIA phos, and do various analyses from there



Version info




###Initialise

Packages
```{r}
library(ksea)
library(reshape2)
library(multcomp)
library(rlist)
source("..\\scripts/essential_visualisations_JvG.R")
source("..\\scripts/boxplots_JvG.R")
source("..\\scripts/enrichment_analysis_JvG.R")
source("..\\scripts/3T3L1_mouse_IR__DF__202104_objects.R")
source("..\\scripts/3T3L1_mouse_IR__DF__202104_functions.R")
```


Load in data
```{r}
load("data/intermediate/phos_3t3/phos_3t3_proc__4.RData")
load("data/intermediate/gsk3i/phos_gsk3i_proc__2.RData")
```




###Add GSK3i data to DIA phos and assess reg, defective
```{r}
####Match to phos_3t3_proc

#Quantified overlap
#Ppeptides: 11264
#Psites: 9294
#Pproteins: 2630
length(which(rownames(phos_3t3_proc)[which(rowSums(is.na(phos_3t3_proc[, 1:52])) < 52)] %in%
        rownames(phos_gsk3i_proc)[which(rowSums(is.na(phos_gsk3i_proc[, 1:8])) < 8)]))
length(unique(phos_3t3_proc$uniprot_site[which(rownames(phos_3t3_proc)[which(rowSums(is.na(phos_3t3_proc[, 1:52])) < 52)] %in%
        rownames(phos_gsk3i_proc)[which(rowSums(is.na(phos_gsk3i_proc[, 1:8])) < 8)])]))
length(unique(phos_3t3_proc$uniprot[which(rownames(phos_3t3_proc)[which(rowSums(is.na(phos_3t3_proc[, 1:52])) < 52)] %in%
        rownames(phos_gsk3i_proc)[which(rowSums(is.na(phos_gsk3i_proc[, 1:8])) < 8)])]))


#Match ppeptides to overall psite activity in GSK3i


#Make phos_gsk3i_proc_site_summary_df
phos_gsk3i_proc_site_summary_df <- data.frame(uniprot_site = unique(phos_gsk3i_proc$uniprot_site))
rownames(phos_gsk3i_proc_site_summary_df) <- phos_gsk3i_proc_site_summary_df$uniprot_site
phos_gsk3i_proc_site_summary_df$num_is_up_ppeptides <- 0
phos_gsk3i_proc_site_summary_df$num_is_down_ppeptides <- 0
phos_gsk3i_proc_site_summary_df$num_downregmotif_ppeptides <- 0
for (i in 1:nrow(phos_gsk3i_proc_site_summary_df)){
  
  phos_gsk3i_proc_site_summary_df[i, c("num_is_up_ppeptides",
                                               "num_is_down_ppeptides",
                                               "num_downregmotif_ppeptides")] <- c(length(which(phos_gsk3i_proc$uniprot_site ==
                                                                                             phos_gsk3i_proc_site_summary_df$uniprot_site[i] &
                                                                                             phos_gsk3i_proc$GSK3i_up == TRUE)),
                                                                              length(which(phos_gsk3i_proc$uniprot_site ==
                                                                                             phos_gsk3i_proc_site_summary_df$uniprot_site[i] &
                                                                                             phos_gsk3i_proc$GSK3i_down == TRUE)),
                                                                              length(which(phos_gsk3i_proc$uniprot_site ==
                                                                                             phos_gsk3i_proc_site_summary_df$uniprot_site[i] &
                                                                                             phos_gsk3i_proc$GSK3i_down_motif == TRUE)))
}

##GSK3i_psite_up/down in 3T3-L1, and downregmotif
#I label a ppeptide as up if its psite has at least one ppeptide up in GSK3i; same for down
phos_3t3_proc$GSK3i_psite_up <- FALSE
phos_3t3_proc$GSK3i_psite_up[which(phos_3t3_proc$uniprot_site %in%
                                                    phos_gsk3i_proc_site_summary_df$uniprot_site[which(phos_gsk3i_proc_site_summary_df$num_is_up_ppeptides > 0)])] <- TRUE
phos_3t3_proc$GSK3i_psite_down <- FALSE
phos_3t3_proc$GSK3i_psite_down[which(phos_3t3_proc$uniprot_site %in%
                                                    phos_gsk3i_proc_site_summary_df$uniprot_site[which(phos_gsk3i_proc_site_summary_df$num_is_down_ppeptides > 0)])] <- TRUE
phos_3t3_proc$GSK3i_psite_downregmotif <- FALSE
phos_3t3_proc$GSK3i_psite_downregmotif[which(phos_3t3_proc$uniprot_site %in%
                                                    phos_gsk3i_proc_site_summary_df$uniprot_site[which(phos_gsk3i_proc_site_summary_df$num_downregmotif_ppeptides > 0)])] <- TRUE

#Numbers: Quantified and GSK3i_psite_up/down
#ppetides: 498 up, 761 down, 477 downregmotif
#psites: 264 up, 396 down, 242 downregmotif
#Proteins: 202 up, 222 down, 167 downregmotif


length(which(phos_3t3_proc$GSK3i_psite_up == TRUE &
               rowSums(is.na(phos_3t3_proc[, 1:52])) < 52))
length(which(phos_3t3_proc$GSK3i_psite_down == TRUE &
               rowSums(is.na(phos_3t3_proc[, 1:52])) < 52))
length(which(phos_3t3_proc$GSK3i_psite_downregmotif == TRUE &
               rowSums(is.na(phos_3t3_proc[, 1:52])) < 52))
length(unique(phos_3t3_proc$uniprot_site[which(phos_3t3_proc$GSK3i_psite_up == TRUE &
               rowSums(is.na(phos_3t3_proc[, 1:52])) < 52)]))
length(unique(phos_3t3_proc$uniprot_site[which(phos_3t3_proc$GSK3i_psite_down == TRUE &
               rowSums(is.na(phos_3t3_proc[, 1:52])) < 52)]))
length(unique(phos_3t3_proc$uniprot_site[which(phos_3t3_proc$GSK3i_psite_downregmotif == TRUE &
               rowSums(is.na(phos_3t3_proc[, 1:52])) < 52)]))
length(unique(phos_3t3_proc$uniprot[which(phos_3t3_proc$GSK3i_psite_up == TRUE &
               rowSums(is.na(phos_3t3_proc[, 1:52])) < 52)]))
length(unique(phos_3t3_proc$uniprot[which(phos_3t3_proc$GSK3i_psite_down == TRUE &
               rowSums(is.na(phos_3t3_proc[, 1:52])) < 52)]))
length(unique(phos_3t3_proc$uniprot[which(phos_3t3_proc$GSK3i_psite_downregmotif == TRUE &
               rowSums(is.na(phos_3t3_proc[, 1:52])) < 52)]))


##GSK3i_site_up/down_insreg_up/down in 3T3-L1
phos_3t3_proc$GSK3i_psite_up_insreg_up <- FALSE
phos_3t3_proc$GSK3i_psite_up_insreg_up[which(phos_3t3_proc$GSK3i_psite_up == TRUE &
                                                        phos_3t3_proc$CTRL_reg == "up")] <- TRUE
phos_3t3_proc$GSK3i_psite_down_insreg_down <- FALSE
phos_3t3_proc$GSK3i_psite_down_insreg_down[which(phos_3t3_proc$GSK3i_psite_down == TRUE &
                                                        phos_3t3_proc$CTRL_reg == "down")] <- TRUE
phos_3t3_proc$GSK3i_psite_downregmotif_insreg_down <- FALSE
phos_3t3_proc$GSK3i_psite_downregmotif_insreg_down[which(phos_3t3_proc$GSK3i_psite_downregmotif == TRUE &
                                                        phos_3t3_proc$CTRL_reg == "down")] <- TRUE

#Numbers: GSK3i_psite_up/down_insreg_up/down in 3T3-L1
#ppetides: 52 up, 99 down, 65 downregmotif
#ppsites: 45 up, 89 down, 57 downregmotif
#Proteins: 41 up, 57 down, 44 downregmotif


length(which(phos_3t3_proc$GSK3i_psite_up_insreg_up == TRUE &
               rowSums(is.na(phos_3t3_proc[, 1:52])) < 52))
length(which(phos_3t3_proc$GSK3i_psite_down_insreg_down == TRUE &
               rowSums(is.na(phos_3t3_proc[, 1:52])) < 52))
length(which(phos_3t3_proc$GSK3i_psite_downregmotif_insreg_down == TRUE &
               rowSums(is.na(phos_3t3_proc[, 1:52])) < 52))
length(unique(phos_3t3_proc$uniprot_site[which(phos_3t3_proc$GSK3i_psite_up_insreg_up == TRUE &
               rowSums(is.na(phos_3t3_proc[, 1:52])) < 52)]))
length(unique(phos_3t3_proc$uniprot_site[which(phos_3t3_proc$GSK3i_psite_down_insreg_down == TRUE &
               rowSums(is.na(phos_3t3_proc[, 1:52])) < 52)]))
length(unique(phos_3t3_proc$uniprot_site[which(phos_3t3_proc$GSK3i_psite_downregmotif_insreg_down == TRUE &
               rowSums(is.na(phos_3t3_proc[, 1:52])) < 52)]))
length(unique(phos_3t3_proc$uniprot[which(phos_3t3_proc$GSK3i_psite_up_insreg_up == TRUE &
               rowSums(is.na(phos_3t3_proc[, 1:52])) < 52)]))
length(unique(phos_3t3_proc$uniprot[which(phos_3t3_proc$GSK3i_psite_down_insreg_down == TRUE &
               rowSums(is.na(phos_3t3_proc[, 1:52])) < 52)]))
length(unique(phos_3t3_proc$uniprot[which(phos_3t3_proc$GSK3i_psite_downregmotif_insreg_down == TRUE &
               rowSums(is.na(phos_3t3_proc[, 1:52])) < 52)]))
```

Defective:
```{r}
##Numbers: Downstream of GSK3 and defective in at least one model
#GSK3i_up_INS_up, GSK3i_down_INS_down, GSK3i_downregmotif_INS_down
#ppeptides: 9 up, 48 down, 29 downregmotif
#psites: 9 up, 44 down, 25 downregmotif
#proteins: 9 up, 29 down, 19 downregmotif

length(which(phos_3t3_proc$GSK3i_psite_up_insreg_up == TRUE &
               phos_3t3_proc$is_defective == TRUE))
length(which(phos_3t3_proc$GSK3i_psite_down_insreg_down == TRUE &
               phos_3t3_proc$is_defective == TRUE))
length(which(phos_3t3_proc$GSK3i_psite_downregmotif_insreg_down == TRUE &
               phos_3t3_proc$is_defective == TRUE))

length(unique(phos_3t3_proc$uniprot_site[which(phos_3t3_proc$GSK3i_psite_up_insreg_up == TRUE &
               phos_3t3_proc$is_defective == TRUE)]))
length(unique(phos_3t3_proc$uniprot_site[which(phos_3t3_proc$GSK3i_psite_down_insreg_down == TRUE &
               phos_3t3_proc$is_defective == TRUE)]))
length(unique(phos_3t3_proc$uniprot_site[which(phos_3t3_proc$GSK3i_psite_downregmotif_insreg_down == TRUE &
               phos_3t3_proc$is_defective == TRUE)]))

length(unique(phos_3t3_proc$uniprot[which(phos_3t3_proc$GSK3i_psite_up_insreg_up == TRUE &
               phos_3t3_proc$is_defective == TRUE)]))
length(unique(phos_3t3_proc$uniprot[which(phos_3t3_proc$GSK3i_psite_down_insreg_down == TRUE &
               phos_3t3_proc$is_defective == TRUE)]))
length(unique(phos_3t3_proc$uniprot[which(phos_3t3_proc$GSK3i_psite_downregmotif_insreg_down == TRUE &
               phos_3t3_proc$is_defective == TRUE)]))


###Defect summary (all downregulated and defective sites)
gsk3_3t3_downdefect_summary_df <- reg_sites_summariser(phos_3t3_proc[which(phos_3t3_proc$GSK3i_psite_downregmotif_insreg_down == TRUE), ])





###Defective sites: Heatmap and output

##Heatmap
fc_sig_heatmap(data = phos_3t3_proc[which(phos_3t3_proc$GSK3i_psite_downregmotif_insreg_down == TRUE &
               phos_3t3_proc$is_defective == TRUE), ],
               fc_cols_w_pvals = c("CTRL_FC",
                           "CI_FC",
                           "DEX_FC",
                           "TNF_FC",
                           "MPQ_FC",
                           "AA_FC"),
               pval_cols = c("t_adj_p_val",
                             "CI_padj",
                             "DEX_padj",
                             "TNF_padj",
                             "MPQ_padj",
                             "AA_padj"),
               order_column = "CTRL_FC",
               is_decreasing = FALSE,
               x_axis_names = c("CTRL",
                                "CI",
                                "DEX",
                                "TNF",
                                "MPQ",
                                "AA")) +
  ggtitle("GSK3 defect sites")
  ggsave("output/images/Gsk3/dia_putative_gsk3_sites_defective_geq1.pdf",
         width = 3,
         height = 5)
  
  
write.csv(phos_3t3_proc[which(phos_3t3_proc$GSK3i_psite_downregmotif_insreg_down == TRUE &
               phos_3t3_proc$is_defective == TRUE), grep("defect",
                                                         colnames(phos_3t3_proc))],
          file = "output/data/gsk3/dia_putative_gsk3_sites_defective_geq1.csv")
```




###KSEA using GSK3i ppome

Using only putative sites with negative FC in CTRL
```{r, fig.show = "hide"}
###Make dataframe for Gsk3i ppome that looks like psp_kinase_substrate
#So that the code works, I add in 2 copies of the same data (it needs there to be two or more kinases)
phos_gsk3i_kinase_substrate <- as.data.frame(cbind("GSK3",
                                rep(unique(phos_gsk3i_proc$uniprot_site[which(phos_gsk3i_proc$GSK3i_down_motif == TRUE)]),
    2)))
colnames(phos_gsk3i_kinase_substrate) <- c("KINASE",
                                           "uniprot_site")
phos_gsk3i_kinase_substrate$KINASE[(0.5*nrow(phos_gsk3i_kinase_substrate) + 1):nrow(phos_gsk3i_kinase_substrate)] <- "GSK3_copy"


####Run enrichment

##INS/BAS
#Standard
set.seed(28)
phos_gsk3i_ksea_INSBAS_output_CTRLnegFC <- ksea_beltrao_enricher_mult_conditions(data = phos_3t3_proc,
                                                                       psp_ks = phos_gsk3i_kinase_substrate,
                                                             conditions = c("CTRL",
                                                                  "CI",
                                                                  "DEX",
                                                                  "TNF",
                                                                  "MPQ",
                                                                  "AA"),
                                                             sub_search_data = phos_3t3_proc[which(phos_3t3_proc$CTRL_FC < 0), ])

#FC Normalised to CTRL
#Make columns
phos_3t3_proc$CI_FC_on_control_FC <- phos_3t3_proc$CI_FC - phos_3t3_proc$CTRL_FC
phos_3t3_proc$DEX_FC_on_control_FC <- phos_3t3_proc$DEX_FC - phos_3t3_proc$CTRL_FC
phos_3t3_proc$TNF_FC_on_control_FC <- phos_3t3_proc$TNF_FC - phos_3t3_proc$CTRL_FC
phos_3t3_proc$MPQ_FC_on_control_FC <- phos_3t3_proc$MPQ_FC - phos_3t3_proc$CTRL_FC
phos_3t3_proc$AA_FC_on_control_FC <- phos_3t3_proc$AA_FC - phos_3t3_proc$CTRL_FC
set.seed(11)
phos_gsk3i_ksea_INSBAS_output_CTRLnegFC_FCnormtoCTRL <- ksea_beltrao_enricher_mult_conditions(data = phos_3t3_proc,
                                                                       psp_ks = phos_gsk3i_kinase_substrate,
                                                             conditions = c("CI_FC_on_control",
                                                                  "DEX_FC_on_control",
                                                                  "TNF_FC_on_control",
                                                                  "MPQ_FC_on_control",
                                                                  "AA_FC_on_control"),
                                                             sub_search_data = phos_3t3_proc[which(phos_3t3_proc$CTRL_FC < 0), ])
save(phos_gsk3i_ksea_INSBAS_output_CTRLnegFC_FCnormtoCTRL,
     file = "output/data/phos_gsk3i_ksea_INSBAS_output_CTRLnegFC_FCnormtoCTRL__1.RData")


##BAS MODEL/CTRL
#Make renamed cols so it will work
phos_3t3_proc$CI_FC_on_control_BAS_FC <- phos_3t3_proc$CI_FC_on_control_BAS
phos_3t3_proc$DEX_FC_on_control_BAS_FC <- phos_3t3_proc$DEX_FC_on_control_BAS
phos_3t3_proc$TNF_FC_on_control_BAS_FC <- phos_3t3_proc$TNF_FC_on_control_BAS
phos_3t3_proc$MPQ_FC_on_control_BAS_FC <- phos_3t3_proc$MPQ_FC_on_control_BAS
phos_3t3_proc$AA_FC_on_control_BAS_FC <- phos_3t3_proc$AA_FC_on_control_BAS
set.seed(85)
phos_gsk3i_ksea_BASMODELCTRL_output_CTRLnegFC <- ksea_beltrao_enricher_mult_conditions(data = phos_3t3_proc,
                                                            psp_ks =  phos_gsk3i_kinase_substrate,
                                                             conditions = c("CI_FC_on_control_BAS",
                                                                  "DEX_FC_on_control_BAS",
                                                                  "TNF_FC_on_control_BAS",
                                                                  "MPQ_FC_on_control_BAS",
                                                                  "AA_FC_on_control_BAS"),
                                                             sub_search_data = phos_3t3_proc[which(phos_3t3_proc$CTRL_FC < 0), ])

##INS MODEL/CTRL
#Make renamed cols so it will work
phos_3t3_proc$CI_FC_on_control_INS_FC <- phos_3t3_proc$CI_FC_on_control_INS
phos_3t3_proc$DEX_FC_on_control_INS_FC <- phos_3t3_proc$DEX_FC_on_control_INS
phos_3t3_proc$TNF_FC_on_control_INS_FC <- phos_3t3_proc$TNF_FC_on_control_INS
phos_3t3_proc$MPQ_FC_on_control_INS_FC <- phos_3t3_proc$MPQ_FC_on_control_INS
phos_3t3_proc$AA_FC_on_control_INS_FC <- phos_3t3_proc$AA_FC_on_control_INS
set.seed(101)
phos_gsk3i_ksea_INSMODELCTRL_output_CTRLnegFC <- ksea_beltrao_enricher_mult_conditions(data = phos_3t3_proc,
                                                             psp_ks = phos_gsk3i_kinase_substrate,
                                                             conditions = c("CI_FC_on_control_INS",
                                                                  "DEX_FC_on_control_INS",
                                                                  "TNF_FC_on_control_INS",
                                                                  "MPQ_FC_on_control_INS",
                                                                  "AA_FC_on_control_INS"),
                                                             sub_search_data = phos_3t3_proc[which(phos_3t3_proc$CTRL_FC < 0), ])


####Plot

###INS/BAS enrichment
phos_gsk3i_ksea_ES_CTRLnegFC_melt <- melt(phos_gsk3i_ksea_INSBAS_output_CTRLnegFC[1, grep("ES",
                                                                       colnames(phos_gsk3i_ksea_INSBAS_output_CTRLnegFC))])
phos_gsk3i_ksea_ES_CTRLnegFC_melt$condition <- factor(gsub("_ES",
                                                    "",
                                                    phos_gsk3i_ksea_ES_CTRLnegFC_melt$variable),
                                               levels = c("CTRL",
                                                          "CI",
                                                          "DEX",
                                                          "TNF",
                                                          "MPQ",
                                                          "AA"))

ggplot(data = phos_gsk3i_ksea_ES_CTRLnegFC_melt,
       aes(x = condition,
           y = value,
           fill = condition)) +
  geom_col() +
  scale_fill_manual(values = SH_control_model_main_colours,
                    name = "Condition") +
  geom_hline(yintercept = 0) +
 comfy_theme() +
  theme(axis.text.x = element_text(colour = "black",
                                   angle = 90,
                                   hjust = 1,
                                   vjust = 0.5),
        axis.line.x = element_blank(),
        legend.position = "none") +
  labs(x = "Condition",
       y = "ES") +
  ggtitle("KSEA using GSK3i putative substrates") 
  ggsave("output/images/Gsk3/KSEA/CTRLnegFC_put_sites/Gsk3i_put_substrates_CTRLnegFC_KSEA.pdf",
         width = 3,
         height = 3)




####CTRL/MODEL enrichment

###No protein normalisation

##Set up
phos_gsk3i_ksea_MODELCTRL_output_CTRLnegFC <- cbind(phos_gsk3i_ksea_BASMODELCTRL_output_CTRLnegFC,
                                          phos_gsk3i_ksea_INSMODELCTRL_output_CTRLnegFC)
phos_gsk3i_ksea_MODELCTRL_output_CTRLnegFC_melt <- cbind(melt(phos_gsk3i_ksea_MODELCTRL_output_CTRLnegFC[1, grep("ES",
                                             colnames(phos_gsk3i_ksea_MODELCTRL_output_CTRLnegFC))]),
                                             melt(phos_gsk3i_ksea_MODELCTRL_output_CTRLnegFC[1, grep("S_pval",
                                             colnames(phos_gsk3i_ksea_MODELCTRL_output_CTRLnegFC))]))
colnames(phos_gsk3i_ksea_MODELCTRL_output_CTRLnegFC_melt) <- c("ES_var",
                                                     "ES",
                                                     "pval_var",
                                                     "pval")
phos_gsk3i_ksea_MODELCTRL_output_CTRLnegFC_melt$condition <- gsub("FC_on_control_",
                                                        "",
                                                        as.character(phos_gsk3i_ksea_MODELCTRL_output_CTRLnegFC_melt$ES_var))
phos_gsk3i_ksea_MODELCTRL_output_CTRLnegFC_melt$condition <- gsub("_ES",
                                                        "",
                                                        phos_gsk3i_ksea_MODELCTRL_output_CTRLnegFC_melt$condition)
phos_gsk3i_ksea_MODELCTRL_output_CTRLnegFC_melt$condition <- factor(phos_gsk3i_ksea_MODELCTRL_output_CTRLnegFC_melt$condition,
                                                          levels = unique(phos_gsk3i_ksea_MODELCTRL_output_CTRLnegFC_melt$condition))
phos_gsk3i_ksea_MODELCTRL_output_CTRLnegFC_melt$model <- 
  factor(sapply(as.character(phos_gsk3i_ksea_MODELCTRL_output_CTRLnegFC_melt$condition),
                function(x) strsplit(x,
                                     "_")[[1]][1]),
         levels = c("CI",
                    "DEX",
                    "TNF",
                    "MPQ",
                    "AA"))


#Plot
ggplot() +
  geom_col(data = phos_gsk3i_ksea_MODELCTRL_output_CTRLnegFC_melt,
           aes(x = condition,
               y = ES,
               colour = model),
           fill = "white",
           size = 0.4,
           width = 0.8) +
  geom_col(data = 
             phos_gsk3i_ksea_MODELCTRL_output_CTRLnegFC_melt[which(phos_gsk3i_ksea_MODELCTRL_output_CTRLnegFC_melt$pval < 0.05), ],
           aes(x = condition,
               y = ES,
               fill = model),
           width = 0.8) +
  geom_hline(yintercept = 0) +
  scale_colour_manual(values = SH_model_main_colours,
                    name = "Model") + 
  scale_fill_manual(values = SH_model_main_colours,
                    name = "Model") +
  comfy_theme() +
  theme(axis.text.x = element_text(colour = "black",
                                   angle = 90,
                                   hjust = 1,
                                   vjust = 0.5),
        axis.line.x = element_blank(),
        legend.position = "none") +
  labs(x = "Condition",
       y = "Enrichment score") +
  ggtitle("Putative GSK3 substrates") 
  ggsave("output/images/Gsk3/KSEA/CTRLnegFC_put_sites/Gsk3i_put_substrates_KSEA_BASINSMODELCTRL_CTRLnegFC.pdf",
         width = 3,
         height = 3)
  
#Export
write.csv(phos_gsk3i_ksea_INSBAS_output_CTRLnegFC,
          file = "output/data/Gsk3/phos_gsk3i_ksea_INSBAS_output_CTRLnegFC.csv")
write.csv(phos_gsk3i_ksea_MODELCTRL_output_CTRLnegFC,
          file = "output/data/Gsk3/phos_gsk3i_ksea_MODELCTRL_output_CTRLnegFC.csv")
write.csv(phos_gsk3i_ksea_INSBAS_output_CTRLnegFC_FCnormtoCTRL,
          file = "output/data/Gsk3/phos_gsk3i_ksea_INSBAS_output_CTRLnegFC_FCnormtoCTRL.csv")
```


###Heatmap of all putative substrates quantified
```{r}
####Only INS/BAS

###Criteria:

#Has a FC in at least 4 conditions, including CTRL
fc_sig_heatmap(data = phos_3t3_proc[which(phos_3t3_proc$GSK3i_psite_downregmotif == TRUE &
                        rowSums(is.na(phos_3t3_proc[, c("CTRL_FC",
                           "CI_FC",
                           "DEX_FC",
                           "TNF_FC",
                           "MPQ_FC",
                           "AA_FC")])) <= 2 &
                        is.na(phos_3t3_proc[, "CTRL_FC"]) == FALSE), ],
               fc_cols_wo_pvals = c("CTRL_FC",
                           "CI_FC",
                           "DEX_FC",
                           "TNF_FC",
                           "MPQ_FC",
                           "AA_FC"),
               order_column = "CTRL_FC",
               is_decreasing = FALSE,
               x_axis_names = c("CTRL",
                                "CI",
                                "DEX",
                                "TNF",
                                "MPQ",
                                "AA"),
               scale_colours_manually = TRUE,
               tile_border_size = 0) +
  ggtitle("Putative GSK3 substrates") 
  ggsave("output/images/Gsk3/all_put_gsk3_subs_3t3l1phos_quantgeq4.pdf",
         height = 14,
         width = 8)
```


####S9 and S21
```{r}
####Site level anovas and dunnetts
###Raw intensity

pGsk3_dunnetts <- list(
  "Gsk3a_S21" = site_level_anova_dunnetts_3t3l1(data = phos_3t3_proc,
                                site = "Gsk3a_Q2NL51_S21_1"),
  
  "Gsk3b_S9" = site_level_anova_dunnetts_3t3l1(data = phos_3t3_proc,
                                site = "Gsk3b_Q9WV60_S9_1"),
  
  "Gsk3a_S21_protnorm" = site_level_anova_dunnetts_3t3l1(data = phos_3t3_proc,
                                site = "Gsk3a_Q2NL51_S21_1",
                                protnorm = TRUE),
  
  "Gsk3b_S9_protnorm" = site_level_anova_dunnetts_3t3l1(data = phos_3t3_proc,
                                site = "Gsk3b_Q9WV60_S9_1",
                                protnorm = TRUE)
)

####Plots
###Gsk3a S21
boxplot_3t3phos(data = phos_3t3_proc,
                                      site = "Gsk3a_Q2NL51_S21_1") 
  ggsave_pdfpng("output/images/Gsk3/reg_sites/Gsk3a_Q2NL51_S21_1_dia",
         width = 2.2,
         height = 1.8)
boxplot_3t3phos(data = phos_3t3_proc,
                                      site = "Gsk3a_Q2NL51_S21_1",
                                      prot_normalised = TRUE) 
  ggsave_pdfpng("output/images/Gsk3/reg_sites/Gsk3a_Q2NL51_S21_1_dia_protnorm",
        width = 2.2,
         height = 1.8)

###Gsk3b S9
boxplot_3t3phos(data = phos_3t3_proc,
                                      site = "Gsk3b_Q9WV60_S9_1") 
  ggsave_pdfpng("output/images/Gsk3/reg_sites/Gsk3b_Q9WV60_S9_1_dia",
        width = 2.2,
         height = 1.8)

boxplot_3t3phos(data = phos_3t3_proc,
                                      site = "Gsk3b_Q9WV60_S9_1",
                                      prot_normalised = TRUE) 
  ggsave_pdfpng("output/images/Gsk3/reg_sites/Gsk3b_Q9WV60_S9_1_dia_protnorm",
       width = 2.2,
         height = 1.8)
  
##Output stats
pGsk3_dunnetts_df <- map(pGsk3_dunnetts, ~c(.$BAS_dunnetts_ps, .$INS_dunnetts_ps)) %>%
  purrr::reduce(rbind) %>% as.data.frame
rownames(pGsk3_dunnetts_df) <- names(pGsk3_dunnetts)
colnames(pGsk3_dunnetts_df) <- cross(list(c("CI", "DEX", "TNF", "MPQ", "AA"), 
                                          c("bas", "ins"))) %>%
  map(~paste(unlist(.), collapse = "_")) %>% unlist

write.csv(pGsk3_dunnetts_df,
          file = "output/data/Gsk3/S9S21/Gsk3_s9s21_dunnetts_ps.csv")
```

###pY
Here we look at the proposed regulatory pYs on Gsk3
```{r}
library(purrr)
Gsk3_pYs <- c("Gsk3a_Q2NL51_Y279_1",
              "Gsk3a_Q2NL51_Y279_2")

#3T3L1
#Gsk3a
gsk3a_3t3_pY_args <- list(
  list(protnorm = FALSE,
       file = "_dia"),
  list(protnorm = TRUE,
       file = "_dia_protnorm")
)
map(gsk3a_3t3_pY_args,
    function(x){
      boxplot_3t3phos(phos_3t3_proc,
                      "Gsk3a_Q2NL51_Y279_1",
                      prot_normalised = x$protnorm)
      ggsave_pdfpng(paste("output/images/Gsk3/reg_sites/Gsk3a_Q2NL51_Y279_1",
                          x$file,
                          sep = ""),
              width = 2.2,
               height = 1.8)
})
```


###Total protein
```{r}
load("data/intermediate/prot_3t3/proteome_post_analysis_3t3_DIA_matched__1.RData")

proteome_post_analysis_3t3_DIA_matched[grep("Gsk3", rownames(proteome_post_analysis_3t3_DIA_matched)), ]

#Gsk3a
boxplot_grouped_3t3_proteome(data = proteome_post_analysis_3t3_DIA_matched,
                                      protein = "Gsk3a_Q2NL51") 
  ggsave_pdfpng("output/images/Gsk3/total_protein/Gsk3a_Q2NL51",
       width = 1.5,
         height = 1.5)
#Gsk3b
boxplot_grouped_3t3_proteome(data = proteome_post_analysis_3t3_DIA_matched,
                                      protein = "Gsk3b_Q9WV60") 
  ggsave_pdfpng("output/images/Gsk3/total_protein/Gsk3b_Q9WV60",
       width = 1.5,
         height = 1.5)
```


###Example defect sites
```{r}
eg_Gsk3_def_sites <- c("Gys1_Q9Z1E4_S641_2",
                       "Kat5_Q8CHK4_S86_2",
                       "Kat5_Q8CHK4_S90_2")

boxplot_3t3phos(phos_3t3_proc,
                site = eg_Gsk3_def_sites,
                facet = TRUE,
                facet_labels = sapply(eg_Gsk3_def_sites, function(x){
                  x_split <- strsplit(x, "_")[[1]]
                  paste(c(x_split[1], " ", x_split[3], " P", x_split[4]), collapse = "")
                }),
                facet_ncol_nrow = c(3, 1),
                siglabels = TRUE)
ggsave_pdfpng("output/images/Gsk3/defect/eg_Gsk3_def_sites_Gys1S641_Kat5S86S90_simpler.pdf",
              width = 6, height = 2)

ggsave(boxplot_all_conditions_3t3_post202104_mult_panel(data = phos_3t3_proc,
                                                 sites = eg_Gsk3_def_sites,
                                                 ncol = 3,
                                                 nrow = 1),
       file = "output/images/Gsk3/defect/eg_Gsk3_def_sites_Gys1S641_Kat5S86S90.pdf",
       width = 6,
       height = 1.3)
```
